### 2022-09-29 #4809 @editorialbot
[PRE REVIEW]: HydroMT: Automated and reproducible model building and analysis
**Submitting author:** @dirkeilander (<a href="http://orcid.org/0000-0002-0951-8418">Dirk Eilander</a>)
**Repository:** https://github.com/Deltares/hydromt
**Branch with paper.md** (empty if default branch): joss_paper
**Version:** v0.5.0
