+++ hydromt/models/model_api.py
@@ -75,11 +75,11 @@ class Model(object, metaclass=ABCMeta):
         data_libs : List[str], optional
             List of data catalog yaml files, by default None
         """
-        from . import ENTRYPOINTS  # load within method to avoid circular imports
+        from . import MODELS  # avoid circular import
 
         self.logger = logger
-        ep = ENTRYPOINTS.get(self._NAME, None)
-        version = ep.distro.version if ep is not None else ""
+        ep = MODELS.eps.get(self._NAME, None)
+        version = ep.distro.version if ep is not None else "NA"
         dist = ep.distro.name if ep is not None else "unknown"
 
         # link to data
