+++ hydromt/__init__.py
@@ -15,15 +15,3 @@ from . import cli, workflows, stats, flw, raster, vector
 from .models import *
 from .io import *
 from .data_catalog import *
-
-# TODO remove after fixed in all plugins
-def __getattr__(name):
-    if name in PLUGINS:
-        ep = ENTRYPOINTS[PLUGINS[name]]
-        plugin_name = ep.module_name.split(".")[0]
-        warnings.warn(
-            f'"hydromt.{name}" will be deprecated, use "{plugin_name}.{name}" instead.',
-            DeprecationWarning,
-        )
-        return model_plugins.load(ep)
-    raise AttributeError(f"module '{__name__}' has no attribute '{name}'")
