+++ hydromt/models/model_plugins.py
@@ -1,14 +1,44 @@
-import itertools
-import entrypoints
+from entrypoints import EntryPoint, Distribution, get_group_all
 import logging
+from typing import Dict, Iterator, List
+from .. import __version__
+from .model_api import Model
 
 logger = logging.getLogger(__name__)
 
+__all__ = ["ModelCatalog"]
 
-def discover(path=None, logger=logger):
-    """Discover hydromt models:
-    - Find 'hydromt.models' entrypoints provided by any Python packages in the
-      environment.
+# local generic models
+LOCAL_EPS = {
+    "grid_model": "hydromt.models.model_grid:GridModel",
+    "lumped_model": "hydromt.models.model_lumped:LumpedModel",
+    "mesh_model": "hydromt.models.model_mesh:MeshModel",
+    "network_model": "hydromt.models.model_network:NetworkModel",
+}
+
+
+def get_general_eps() -> Dict:
+    """Get local hydromt generic model class entrypoints
+
+    Returns
+    -------
+    eps : dict
+        Entrypoints dict
+    """
+    eps = {}
+    distro = Distribution("hydromt", __version__)
+    for name, epstr in LOCAL_EPS.items():
+        eps[name] = EntryPoint.from_string(epstr, name, distro)
+    return eps
+
+
+def _discover(path=None) -> List:
+    """Discover drivers via entrypoints"""
+    return get_group_all("hydromt.models", path=path)
+
+
+def get_plugin_eps(path=None, logger=logger) -> Dict:
+    """Discover hydromt model plugins based on 'hydromt.models' entrypoints
 
     Parameters
     ----------
@@ -20,22 +50,13 @@ def discover(path=None, logger=logger):
     eps : dict
         Entrypoints dict
     """
-    # Discover drivers via entrypoints.
-    group = entrypoints.get_group_named("hydromt.models", path=path)
-    group_all = entrypoints.get_group_all("hydromt.models", path=path)
-    if len(group_all) != len(group):
-        # There are some name collisions. Let's go digging for them.
-        for name, matches in itertools.groupby(group_all, lambda ep: ep.name):
-            matches = list(matches)
-            if len(matches) != 1:
-                winner = group[name]
-                logger.debug(
-                    f"There are {len(matches)} 'hydromt.models' entrypoints for the name {name}: {matches}."
-                    f"The match {winner} is selected."
-                )
-
     eps = {}
-    for name, ep in group.items():
+    for ep in _discover():
+        name = ep.name
+        if name in eps or name in LOCAL_EPS:
+            plugin = f"{ep.module_name}.{ep.object_name}"
+            logger.warning(f"Duplicated model plugin '{name}'; skipping {plugin}")
+            continue
         logger.debug(
             f"Discovered model plugin '{name} = {ep.module_name}.{ep.object_name}' ({ep.distro.version})"
         )
@@ -43,7 +64,7 @@ def discover(path=None, logger=logger):
     return eps
 
 
-def load(ep, module=None, logger=logger):
+def load(ep, logger=logger) -> Model:
     """Load entrypoint and return plugin model class
 
     Parameters
@@ -56,14 +77,83 @@ def load(ep, module=None, logger=logger):
     model_class : Model
         plugin model class
     """
+    _str = f"{ep.name} = {ep.module_name}.{ep.object_name}"
     try:
         # plugins[ep.name] = ep.load()
         model_class = ep.load()
-        if module is not None:
-            setattr(module, model_class.__name__, model_class)
-        logger.debug(
-            f"Loaded model plugin '{ep.name} = {ep.module_name}.{ep.object_name}' ({ep.distro.version})"
-        )
+        if not issubclass(model_class, Model):
[…]
