### 2022-09-30 01e9db6
use load method to get class
