+++ envs/hydromt-test.yml
@@ -23,8 +23,9 @@ dependencies:
   - pyflwdir>=0.5.4
   - pygeos>=0.8
   - pytest # tests
-  - pytest-cov # tests
   - pytest-benchmark # tests
+  - pytest-cov # tests
+  - pytest-mock # tests
   - rasterio
   - requests
   - rioxarray
