+++ hydromt/geodataset.py
@@ -1,0 +1,438 @@
+#!/usr/bin/env python
+# -*- coding: utf-8 -*-
+""""""
+#%%
+import numpy as np
+import xarray as xr
+import pandas as pd
+import geopandas as gpd
+import geopandas.array as geoarray
+from shapely.geometry import box
+from shapely.geometry.base import BaseGeometry
+import pyproj
+import logging
+
+from hydromt import gis_utils, raster
+
+logger = logging.getLogger(__name__)
+
+#%%
+class GeoBase(raster.XGeoBase):
+    def __init__(self, xarray_obj):
+        super(GeoBase, self).__init__(xarray_obj)
+        self._geometry = None  # placeholder
+
+    @property
+    def _all_names(self):
+        names = [n for n in self._obj.coords]
+        if isinstance(self._obj, xr.Dataset):
+            names = names + [n for n in self._obj.data_vars]
+        return names
+
+    @property
+    def _geom_names(self):
+        names = []
+        for name in self._all_names:
+            if self._obj[name].ndim == 1 and isinstance(
+                self._obj[name][0].values.item(), BaseGeometry
+            ):
+                names.append(name)
+        return names
+
+    def _discover_xy(self, x_name=None, y_name=None, index_dim=None):
+        """Set the spatial and index dimensions of the object.
+
+        Arguments
+        ----------
+        x_name, y_name, index_dim: str, optional
+            The name of the x, y and index dimensions.
+        """
+        # infer x dim
+        if x_name is None:
+            for name in raster.XDIMS:
+                if name in self._obj:
+                    dim0 = (
+                        index_dim if index_dim is not None else self._obj[name].dims[0]
+                    )
+                    if self._obj[name].ndim == 1 and self._obj[name].dims[0] == dim0:
+                        x_name = name
+                        index_dim = dim0
+                        break
+        # infer y dim
+        if y_name is None and x_name is not None:
+            for name in raster.YDIMS:
+                if name in self._obj:
+                    if self._obj[name].dims[0] == index_dim:
+                        y_name = name
+                        break
+        if (
+            x_name is not None
+            and y_name is not None
+            and self._obj[x_name].ndim == 1
+            and self._obj[x_name].dims == self._obj[y_name].dims
+        ):
+            self.set_attrs(x_name=x_name)
+            self.set_attrs(y_name=y_name)
+            self.set_attrs(index_dim=index_dim)
+        else:
+            self.attrs.pop("x_name", None)
+            self.attrs.pop("y_name", None)
+            self.attrs.pop("index_dim", None)
+
+    def _discover_geom(self, geom_name=None, index_dim=None):
+        # infer geom dim
+        if geom_name is None:
+            names = self._geom_names
+            if index_dim is not None:
+                names = [name for name in names if self._obj[name].dims[0] == index_dim]
+            if "geometry" in names:
+                geom_name = "geometry"
+            elif len(names) >= 1:
+                geom_name = names[0]
+        if (
+            geom_name is not None
+            and geom_name in self._obj
+            and len(self._obj[geom_name].dims) == 1
+        ):
+            self.set_attrs(geom_name=geom_name)
+            self.set_attrs(index_dim=self._obj[geom_name].dims[0])
+        else:
+            self.attrs.pop("geom_name", None)
+            self.attrs.pop("index_dim", None)
+
+    def set_spatial_dims(
+        self, geom_name=None, x_name=None, y_name=None, index_dim=None
+    ):
[…]
