+++ hydromt/gis_utils.py
@@ -197,11 +197,11 @@ def nearest(
     return gdf2.index.values[idx], dst
 
 
-def filter_gdf(gdf, geom=None, bbox=None, predicate="intersects"):
+def filter_gdf(gdf, geom=None, bbox=None, crs=4326, predicate="intersects"):
     """Filter GeoDataFrame geometries based on geometry mask or bounding box."""
     gtypes = (gpd.GeoDataFrame, gpd.GeoSeries, BaseGeometry)
     if bbox is not None and geom is None:
-        geom = box(*bbox)
+        geom = gpd.GeoSeries([box(*bbox)], crs=crs)
     elif geom is not None and not isinstance(geom, gtypes):
         raise ValueError(
             f"Unknown geometry mask type {type(geom).__name__}. "
