### 2022-10-05 1e72919
Remove shapely pin + update dependencies #176
