### 2022-10-05 59ce8b3
Remove shapely pin + update dependencies #176
