+++ hydromt/workflows/basin_mask.py
@@ -8,7 +8,6 @@ from pathlib import Path
 import numpy as np
 import geopandas as gpd
 from shapely.geometry import box
-from sklearn.neighbors import VALID_METRICS
 import xarray as xr
 import logging
 import warnings
@@ -313,7 +312,7 @@ def get_basin_geometry(
             if "basid" not in gdf_bas.columns:
                 raise ValueError("Basin geometries does not have 'basid' column.")
         if gdf_bas.crs != ds.raster.crs:
-            logger.warn("Basin geometries CRS does not match the input raster CRS.")
+            logger.warning("Basin geometries CRS does not match the input raster CRS.")
             gdf_bas = gdf_bas.to_crs(ds.raster.crs)
 
     ## BASINS
@@ -322,7 +321,7 @@ def get_basin_geometry(
         if basins_name not in ds:
             if gdf_bas is not None:
                 gdf_bas = None
-                logger.warn(
+                logger.warning(
                     "Basin geometries ignored as no corresponding basin map is provided."
                 )
             _check_size(ds, logger)
@@ -383,7 +382,7 @@ def get_basin_geometry(
                     total_bounds = gdf_bas.total_bounds
                 ds = ds[dvars].raster.clip_bbox(total_bounds, buffer=0)
             elif np.any(gdf_match):
-                logger.warn("No matching basin IDs found in basin geometries.")
+                logger.warning("No matching basin IDs found in basin geometries.")
         # get full basin mask and use this mask ds_basin incl flow direction raster
         _mask = np.isin(ds[basins_name], basid)
         if not np.any(_mask > 0):
