+++ tests/test_basin_mask.py
@@ -112,6 +112,7 @@ def test_check_size(caplog):
         nodata=-1,
         name="test",
         crs=4326,
+        lazy=True,  # create lazy dask array instead of numpy array
     )
     _check_size(test_raster)
     assert "Loading very large spatial domain to derive a subbasin. "
@@ -234,13 +235,9 @@ def test_basin(caplog):
     )
     assert gdf_bas.index.size == 180
 
-    with pytest.warns(DeprecationWarning) as record:
+    msg = 'kind="outlets" has been deprecated, use outlets=True in combination with kind="basin" or kind="interbasin" instead.'
+    with pytest.warns(DeprecationWarning, match=msg) as record:
         gdf_bas, gdf_out = get_basin_geometry(ds, kind="outlet")
-    assert len(record) == 1
-    assert (
-        record[0].message.args[0]
-        == 'kind="outlets" has been deprecated, use outlets=True in combination with kind="basin" or kind="interbasin" instead.'
-    )
 
     with pytest.raises(ValueError):
         gdf_bas, gdf_out = get_basin_geometry(ds, kind="watershed")
