### 2022-10-06 94351af
update model docstrings
