+++ hydromt/models/model_lumped.py
@@ -146,7 +146,7 @@ class LumpedModel(LumpedMixin, Model):
         data_libs: List[str] = None,
         logger=logger,
     ):
-        # Initialize the submodel class Lumped Model for lumped models in HydroMT
+        """Initialize a LumpedModel for lumped and semi-distributed models."""
         super().__init__(
             root=root,
             mode=mode,
