+++ docs/user_guide/data_overview.rst
@@ -1,4 +1,4 @@
-.. _get_overview:
+.. _get_data:
 
 Overview data
 =============
