### 2022-10-07 2274239
Added get_region function in api.py and wrote unit tests for this function, also moved other functions from api.py to test_api.py #207
