### 2022-10-07 a9eada2
Removed api.py functions from test_cli.py #207
