+++ tests/test_cli.py
@@ -5,8 +5,6 @@ from click.testing import CliRunner
 import numpy as np
 from hydromt import __version__
 from hydromt.cli.main import main as hydromt_cli
-from hydromt.cli import api as hydromt_api
-from hydromt.models import ENTRYPOINTS
 
 
 def test_cli(tmpdir):
@@ -57,24 +55,3 @@ def test_cli(tmpdir):
     )
     with pytest.raises(NotImplementedError):
         raise r.exception
-
-
-def test_api_datasets():
-    # datasets
-    assert "artifact_data" in hydromt_api.get_predifined_catalogs()
-    datasets = hydromt_api.get_datasets("artifact_data")
-    assert isinstance(datasets, dict)
-    assert isinstance(datasets["RasterDatasetSource"], list)
-
-
-@pytest.mark.skipif("lumped_model" not in ENTRYPOINTS, reason="HydroMT not installed.")
-def test_api_model_components():
-    # models
-    components = hydromt_api.get_model_components(
-        "lumped_model", component_types=["write"]
-    )
-    name = "write_response_units"
-    assert name in components
-    assert np.all([k.startswith("write") for k in components])
-    keys = ["doc", "required", "optional", "kwargs"]
-    assert np.all([k in components[name] for k in keys])
