### 2022-10-07 a56078e
Changed datacatalog to use artifact_data #207
