+++ tests/test_api.py
@@ -5,10 +5,9 @@ import numpy as np
 from hydromt.cli.api import *
 
 
-def test_get_region():
-    data_libs = "deltares_data"
-    region = {"basin": [6.338331593204473, 52.521107365953334]}  # vecht basin
-    region_geom = get_region(region=region, data_libs=data_libs)
+def test_get_region():    
+    region = {"basin": [12.111195, 46.088866]}  # Piave basin
+    region_geom = get_region(region=region)
     assert isinstance(region_geom, str)
     try:
         region_json = json.loads(region_geom)
@@ -23,8 +22,8 @@ def test_get_region():
         match="Only basin, subbasin, and interbasin are accepted region definitions",
     ):
         get_region(
-            region={"bbox": [4.416504, 51.597548, 6.256714, 52.281602]},
-            data_libs=data_libs,
+            region={"bbox": [4.416504, 51.597548, 6.256714, 52.281602]}
+            
         )
 
 
