### 2022-10-10 1daad15
Bugfix in ep.load if an exception error is found
