+++ hydromt/models/model_plugins.py
@@ -84,76 +84,12 @@ def load(ep, logger=logger) -> Model:
     try:
         # plugins[ep.name] = ep.load()
         model_class = ep.load()
-        if not issubclass(model_class, Model):
-            raise ValueError(f"Model plugin type not recognized '{_str}'")
-        logger.debug(f"Loaded model plugin {_str}")
-    except (ModuleNotFoundError, AttributeError) as err:
-        raise ImportError(f"Error while loading model plugin '{_str}' ({err})")
-    return model_class
-
-
-class ModelCatalog:
-    def __init__(self):
-        self._eps = {}  # entrypoints
-        self._cls = {}  # classes
-        self._plugins = []  # names of plugins
-        self._general = []  # names of local model classes
-
-    @property
-    def eps(self) -> Dict:
-        """Returns dictionary with available model entrypoints."""
-        if len(self._eps) == 0:
-            self.plugins  # discover plugins
-            self.generic  # get generic local model classes
-        return self._eps
-
-    @property
-    def cls(self) -> Dict:
-        """Returns dictionary with available model classes."""
-        if len(self._cls) != len(self.eps):
-            for name in self.eps:
-                if name not in self._cls:
-                    self._cls[name] = load(self.eps[name])
-        return self._cls
-
-    @property
-    def plugins(self) -> List:
-        """Returns list with names of model plugins"""
-        if len(self._plugins) == 0:
-            eps = get_plugin_eps()
-            self._plugins = list(eps.keys())
-            self._eps.update(**eps)
-        return self._plugins
-
-    @property
-    def generic(self) -> List:
-        """Returns list with names of generic models"""
-        if len(self._general) == 0:
-            eps = get_general_eps()
-            self._general = list(eps.keys())
-            self._eps.update(**eps)
-        return self._general
-
-    def load(self, name) -> Model:
-        """Returns model class"""
-        if name not in self._cls:
-            self._cls[name] = load(self[name])
-        return self._cls[name]
-
-    def __str__(self):
-        plugins = "".join(
-            [
-                f" - {name} ({self.eps[name].distro.name} {self.eps[name].distro.version})\n"
-                for name in self.plugins
-            ]
+        if module is not None:
+            setattr(module, model_class.__name__, model_class)
+        logger.debug(
+            f"Loaded model plugin '{ep.name} = {ep.module_name}.{ep.object_name}' ({ep.distro.version})"
         )
-        generic = "".join([f" - {name}\n" for name in self.generic])
-        return f"model plugins:\n{plugins}generic models (hydromt {__version__}):\n{generic}"
-
-    def __getitem__(self, name) -> Model:
-        if name not in self.eps:
-            raise ValueError(f"Unknown model {name}; select from {self.eps.keys()}")
-        return self._eps[name]
-
-    def __iter__(self) -> Iterator:
-        return iter(self.eps)
+        return model_class
+    except (ModuleNotFoundError, AttributeError) as err:
+        logger.exception(f"Error while loading entrypoint {ep.name}: {str(err)}")
+        return None
