### 2022-10-10 8749ade
test loading general models w/ and w/o xugrid
