### 2022-10-10 c089529
test loading general models w/ and w/o xugrid
