### 2022-10-10 7c25dab
fix mistake replacing import MeshModel by MODELS.load()
