### 2022-10-12 de7e638
draft of test_river.py
