### 2022-10-19 60bf671
add example to toctree in docs
