+++ docs/user_guide/intro.rst
@@ -115,7 +115,7 @@ Region                          Argument of the *build* and *clip* CLI methods t
    data_conventions.rst
    ../_examples/reading_raster_data.ipynb
    ../_examples/reading_vector_data.ipynb
-   ../_examples/export_data.ipynb
+   ../_examples/reading_point_data.ipynb
 
 .. toctree::
    :maxdepth: 2
