### 2022-10-19 4cfed05
add example to toctree in docs
