+++ docs/conf.py
@@ -373,3 +373,4 @@ nbsphinx_prolog = r"""
             <a href="https://mybinder.org/v2/gh/Deltares/hydromt/main?urlpath=lab/tree/examples/{{ docname|e }}" target="_blank" rel="noopener noreferrer"><img alt="Binder badge" src="https://mybinder.org/badge_logo.svg"></a>
         </div>
 """
+nbsphinx_allow_errors = True
