### 2022-10-19 ff4f724
remove data catalog files from branch
