+++ data/catalogs/deltares_data_geo_dar.yml
@@ -1,30 +1,0 @@
-geo_dar_grand_v13_issues:
-  crs: 
-  data_type: GeoDataFrame
-  driver: vector_table
-  kwargs:
-    x_dim: ??
-    y_dim: ??
-  meta:
-    category: hydrography
-    notes: indication of issues of the GRanD v 1.3 data 
-    paper_doi: 10.1890/100125, 10.5194/essd-14-1869-2022
-    paper_ref: Lehner et al. (2011), Wang et al. (2022)
-    source_url: https://zenodo.org/record/6163413#.Yue_H2NByUl
-  path: hydrography\geo_dar\GRanD_v13_issues.csv
-
-geo_dar_wada_et_al_2017_harmonized:
-  crs: 
-  data_type: GeoDataFrame
-  driver: vector_table
-  kwargs:
-    x_dim: ??
-    y_dim: ??
-    encoding: ??
-  meta:
-    category: hydrography
-    notes:  renaming and units might require some revision, revised capacities and coordinates of Wada et al. (2017)
-    paper_doi: 10.1007/s10712-016-9399-6, 10.5194/essd-14-1869-2022
-    paper_ref: Wada et al. (2017), Wang et al. (2022)
-    source_url: https://zenodo.org/record/6163413#.Yue_H2NByUl
-  path: hydrography\geo_dar\Wada_et_al_2017_harmonized.csv
