### 2022-10-19 b9965e4
remove data catalog files from branch
