### 2022-10-24 bbfe27d
added export_data notebook fixes #218
