### 2022-10-24 9222303
update hydromt icon closes #215
