+++ hydromt/__init__.py
@@ -1,7 +1,7 @@
 """HydroMT: Automated and reproducible model building and analysis"""
 
 # version number without 'v' at start
-__version__ = "0.5.1.dev"
+__version__ = "0.6.0"
 
 import warnings
 
