+++ docs/changelog.rst
@@ -6,8 +6,8 @@ All notable changes to this project will be documented in this page.
 The format is based on `Keep a Changelog`_, and this project adheres to
 `Semantic Versioning`_.
 
-unreleased
-==========
+v0.6.0 (24 October 2022)
+========================
 
 Added
 -----
