+++ docs/getting_started/examples_index.rst
@@ -17,6 +17,11 @@ Reading data:
 * `Reading raster data <../_examples/reading_raster_data.ipynb>`_
 * `Reading vector data <../_examples/reading_vector_data.ipynb>`_
 
+Working with models:
+********************
+
+* `Working with models in python <../_examples/working_with_models.ipynb>`_
+
 GIS operations:
 ***************
 
