+++ docs/getting_started/examples_index.rst
@@ -18,6 +18,11 @@ Data and DataCatalog:
 * `Reading vector data <../_examples/reading_vector_data.ipynb>`_
 * `Exporting data <../examples/export_data.ipynb>`_
 
+Working with models:
+********************
+
+* `Working with models in python <../_examples/working_with_models.ipynb>`_
+
 GIS operations:
 ***************
 
