### 2022-10-25 9614af3
more detailed notebook install
