### 2022-10-25 3598203
replace rasterio.CRS by pyproj.CRS object
