+++ hydromt/raster.py
@@ -18,7 +18,7 @@ import geopandas as gpd
 import xarray as xr
 import dask
 from affine import Affine
-from rasterio.crs import CRS
+from pyproj import CRS
 import rasterio.warp
 import rasterio.fill
 from rasterio import features
@@ -162,7 +162,8 @@ class XGeoBase(object):
         # create new coordinate with attributes in which to save x_dim, y_dim and crs.
         # other spatial properties are always calculated on the fly to ensure consistency with data
         if GEO_MAP_COORD not in self._obj.coords:
-            self._obj.coords[GEO_MAP_COORD] = xr.Variable((), 1)
+            # zero is used by rioxarray
+            self._obj.coords[GEO_MAP_COORD] = xr.Variable((), 0)
 
     @property
     def attrs(self):
@@ -180,7 +181,13 @@ class XGeoBase(object):
     @property
     def crs(self):
         """Return Coordinate Reference System as :py:meth:`pyproj.CRS` object."""
-        return self._obj.rio.crs
+        if "epsg" in self.attrs:
+            crs = CRS.from_epsg(self.attrs["epsg"])
+        elif "crs_wkt" in self.attrs:
+            crs = CRS.from_wkt(self.attrs["crs_wkt"])
+        else:
+            crs = self.set_crs()
+        return crs
 
     @property
     def x_dim(self):
@@ -248,42 +255,53 @@ class XGeoBase(object):
         if check_x == False or check_y == False:
             raise ValueError("raster only applies to regular grids")
 
-    def set_crs(self, input_crs=None):
+    def set_crs(self, input_crs=None) -> CRS:
         """Set the Coordinate Reference System.
 
         Arguments
         ----------
         input_crs: int, dict, or str, optional
-            Coordinate Reference System. Accepts EPSG codes (int or str); proj (str or dict)
+            Coordinate Reference System. Anything accepted by :py:meth:`pyproj.CRS.from_user_input`
+
+        Returns
+        -------
+        crs: CRS object
         """
         crs_names = ["crs_wkt", "crs", "epsg"]
-        names = list(self._obj.coords.keys())
-        if isinstance(self._obj, xr.Dataset):
-            names = names + self.vars
         # user defined
         if input_crs is not None:
             input_crs = CRS.from_user_input(input_crs)
         # look in grid_mapping and data variable attributes
         else:
-            for name in crs_names:
-                # check default > GEO_MAP_COORDS attrs
-                crs = self._obj.coords[GEO_MAP_COORD].attrs.get(name, None)
-                if crs is None:  # global attrs
-                    crs = self._obj.attrs.pop(name, None)
-                for var in names:  # data var and coords attrs
-                    if name in self._obj[var].attrs:
-                        crs = self._obj[var].attrs.pop(name)
-                        break
-                if crs is not None:
-                    # avoid Warning 1: +init=epsg:XXXX syntax is deprecated
-                    crs = crs.strip("+init=") if isinstance(crs, str) else crs
-                    try:
-                        input_crs = CRS.from_user_input(crs)
-                        break
-                    except:
-                        pass
+            names = [GEO_MAP_COORD, None]
+            names.extend([k for k in self._obj.coords.keys() if k != GEO_MAP_COORD])
+            if isinstance(self._obj, xr.Dataset):
+                names.extend(self.vars)
+            # check GEO_MAP_COORD attrs first, then global attrs, then other coords and variables
+            for var in names:
+                if input_crs is not None:
+                    break
+                for name in crs_names:
+                    attrs = self._obj[var].attrs if var else self._obj.attrs
+                    if name in attrs:
[…]
