### 2022-10-25 #4809 @elbeejay
:wave: @giswqs, @jonschwenk, @JannisHoch, @jvdkwast, @amoodie, @jeffskwang, @kratzert, @mcflugen, I am reaching out to y'all as potential reviewers for this submission to JOSS titled: "_HydroMT: Automated and reproducible model building and analysis_". While you all are experts at this intersection of Python (specifically modular, reproducible, and user-friendly packaging) and hydrology, I also appreciate that you may not have the time to take on a review. If you do have the time to conduct a review for JOSS however, below is some additional information about the process.

At JOSS we do open checklist-driven reviews; peer-review criteria can be viewed [here](https://joss.readthedocs.io/en/latest/review_criteria.html). This issue is a "pre-review issue" which we use to find peer-reviewers. Once 3 reviewers are found, we will officially start the review in a dedicated GitHub issue. At present we are asking reviewers to complete reviews in 6 weeks, although this can be extended if needed. If you are not able to review but can recommend someone else, please mention them here (in this case please mention their GitHub handle without the "@" symbol).

If you are interested, please take a look at the journal's [conflict of interest policy](https://joss.readthedocs.io/en/latest/reviewer_guidelines.html?#joss-conflict-of-interest-policy) to ensure you do not have a conflict before agreeing to review this submission.

Thanks again for taking a moment to consider this, feel free to contact me if you have any questions about the JOSS review process. Please do not feel any pressure to accept this review request if you do not have the time or do not feel comfortable reviewing this software package, we appreciate and respect our peer-reviewers' time. If you cannot serve as a reviewer at this time but have a peer, collaborator, student, or colleague who might be available and would be a good fit for this submission, please let me know! This submission in particular seems appropriate to have a student conduct some or all of the peer-review process, as one of the target audiences for this package are students, as well as those with little to no Python knowledge beyond installation. So if interested we can figure out how to set up a "co-review" with a more junior or inexperienced member of the community.

Thanks,
Jay
