### 2022-10-25 d39498d
update model sections in user guide and api
