+++ docs/getting_started/examples_index.rst
@@ -20,6 +20,7 @@ Reading data:
 Working with models:
 ********************
 
+* `Working with models from CLI <../_examples/working_with_models_basics.ipynb>`_
 * `Working with models in python <../_examples/working_with_models.ipynb>`_
 
 GIS operations:
