### 2022-10-25 c499ed1
update model sections in user guide and api
