+++ docs/getting_started/examples_index.rst
@@ -21,6 +21,7 @@ Data and DataCatalog:
 Working with models:
 ********************
 
+* `Working with models from CLI <../_examples/working_with_models_basics.ipynb>`_
 * `Working with models in python <../_examples/working_with_models.ipynb>`_
 
 GIS operations:
