### 2022-10-26 e1c4eae
add flow direction example #220
