+++ docs/user_guide/intro.rst
@@ -135,5 +135,6 @@ Region                          Argument of the *build* and *clip* CLI methods t
 
    methods_main.rst
    methods_stats.rst
-   ../_examples/delineate_basin.ipynb
    ../_examples/working_with_raster.ipynb
+   ../_examples/working_with_flow_directions.ipynb
+   ../_examples/delineate_basin.ipynb
