+++ envs/hydromt-dev.yml
@@ -33,7 +33,6 @@ dependencies:
   - pygeos>=0.8
   - pyproj
   - pytest # tests
-  - pytest-benchmark # tests
   - pytest-cov # tests
   - pytest-mock # tests
   - python>=3.7
