+++ docs/getting_started/examples_index.rst
@@ -16,7 +16,7 @@ Data and DataCatalog:
 
 * `Reading raster data <../_examples/reading_raster_data.ipynb>`_
 * `Reading vector data <../_examples/reading_vector_data.ipynb>`_
-* `Reading point location data <../_examples/reading_point_data.ipynb>`_
+* `Reading geospatial point time-series <../_examples/reading_point_data.ipynb>`_
 
 GIS operations:
 ***************
