+++ docs/user_guide/intro.rst
@@ -131,6 +131,7 @@ Region                          Argument of the *build* and *clip* CLI methods t
    model_region.rst
    ../_examples/working_with_models_basics.ipynb
    ../_examples/working_with_models.ipynb
+   ../_examples/delineate_basin.ipynb
 
 .. toctree::
    :maxdepth: 2
@@ -141,4 +142,3 @@ Region                          Argument of the *build* and *clip* CLI methods t
    methods_stats.rst
    ../_examples/working_with_raster.ipynb
    ../_examples/working_with_flow_directions.ipynb
-   ../_examples/delineate_basin.ipynb
