+++ docs/getting_started/examples_index.rst
@@ -16,11 +16,18 @@ Data and DataCatalog:
 
 * `Reading raster data <../_examples/reading_raster_data.ipynb>`_
 * `Reading vector data <../_examples/reading_vector_data.ipynb>`_
-* `Reading geospatial point time-series <../_examples/reading_point_data.ipynb>`_
+* `Reading geospatial point time-series data <../_examples/reading_point_data.ipynb>`_
+* `Exporting data <../examples/export_data.ipynb>`_
+
+Working with models:
+********************
+
+* `Working with models from CLI <../_examples/working_with_models_basics.ipynb>`_
+* `Working with models in python <../_examples/working_with_models.ipynb>`_
+* `Define hydrological model regions <../_examples/delineate_basin.ipynb>`_
 
 GIS operations:
 ***************
 
 * `Working with raster data <../_examples/working_with_raster.ipynb>`_
 * `Working with flow direction data <../_examples/working_with_flow_directions.ipynb>`_
-* `Delineate basins <../_examples/delineate_basin.ipynb>`_
