### 2022-10-26 88eb5f0
simplify binder using latest release
