+++ environment.yml
@@ -1,0 +1,14 @@
+# minimal environment to install the latest release of hydromt and run the examples
+# used for binder
+name: hydromt
+
+channels:
+  - conda-forge
+
+dependencies:
+  - cartopy
+  - descartes
+  - jupyter
+  - hydromt>=0.6
+  - matplotlib
+  - python=3.10
