### 2022-10-26 1043d3e
simplify binder using latest release
