+++ binder/environment.yml
@@ -1,37 +1,0 @@
-# minimal environment to install hydromt and run examples
-name: hydromt
-
-channels:
-  - conda-forge
-
-dependencies:
-  - affine
-  - black
-  - bottleneck
-  - cartopy # to run examples
-  - click
-  - dask
-  - descartes # to run examples
-  - entrypoints
-  - flit>=3.4
-  - gdal>=3.1
-  - geopandas>=0.10
-  - jupyter # to run examples
-  - netcdf4
-  - matplotlib # to run examples
-  - numba
-  - numpy
-  - openpyxl
-  - pandas
-  - pcraster # optional
-  - pip
-  - pyflwdir>=0.5.4
-  - pygeos>=0.8
-  - python>=3.7
-  - rasterio
-  - requests
-  - rioxarray
-  - scipy
-  - toml
-  - xarray
-  - zarr
