### 2022-10-27 57dec8d
Finalise prep_data_catalog.ipynb
