### 2022-10-27 ae1bc67
Finalise prep_data_catalog.ipynb
