### 2022-10-27 788615c
small notebook typo + use pprint for docs rendering
