+++ docs/getting_started/examples_index.rst
@@ -18,7 +18,7 @@ Data and DataCatalog:
 * `Reading vector data <../_examples/reading_vector_data.ipynb>`_
 * `Reading geospatial point time-series data <../_examples/reading_point_data.ipynb>`_
 * `Preparing a data catalog <../_examples/prep_data_catalog.ipynb>`_
-* `Exporting data <../examples/export_data.ipynb>`_
+* `Exporting data <../_examples/export_data.ipynb>`_
 
 Working with models:
 ********************
