### 2022-10-27 704248c
Fix path in setting up examples
