### 2022-10-27 33addea
add check shape and debruin
