### 2022-10-28 4467e2d
fix consistent read/write mode
