+++ hydromt/models/model_api.py
@@ -3,7 +3,7 @@
 
 from abc import ABCMeta
 import os, glob
-from os.path import join, isdir, isfile, abspath, dirname, basename
+from os.path import join, isdir, isfile, abspath, dirname, basename, isabs
 import xarray as xr
 import numpy as np
 import geopandas as gpd
@@ -325,7 +325,7 @@ class Model(object, metaclass=ABCMeta):
             For a complete overview of all region options,
             see :py:function:~hydromt.workflows.basin_mask.parse_region
         hydrography_fn : str
-            Name of data source for basemap parameters.
+            Name of data source for hydrography data.
             FIXME describe data requirements
         basin_index_fn : str
             Name of data source with basin (bounding box) geometries associated with
@@ -392,6 +392,16 @@ class Model(object, metaclass=ABCMeta):
         if self._root is None:
             raise ValueError("Root unknown, use set_root method")
         return self._root
+
+    @property
+    def _assert_write_mode(self):
+        if not self._write:
+            raise IOError("Model opened in read-only mode")
+
+    @property
+    def _assert_read_mode(self):
+        if not self._read:
+            raise IOError("Model opened in write-only mode")
 
     def set_root(self, root: Optional[str], mode: Optional[str] = "w"):
         """Initialize the model root.
@@ -524,8 +534,9 @@ class Model(object, metaclass=ABCMeta):
     @property
     def config(self) -> Dict[str, Union[Dict, str]]:
         """Model configuration. Returns a (nested) dictionary"""
+        # initialize default config if in write-mode
         if not self._config:
-            self.read_config()  # initialize default config
+            self.read_config()
         return self._config
 
     def set_config(self, *args):
@@ -645,8 +656,7 @@ class Model(object, metaclass=ABCMeta):
         self, config_name: Optional[str] = None, config_root: Optional[str] = None
     ):
         """Write config to <root/config_fn>"""
-        if not self._write:
-            raise IOError("Model opened in read-only mode")
+        self._assert_write_mode
         if config_name is not None:
             self._config_fn = config_name
         elif self._config_fn is None:
@@ -668,9 +678,8 @@ class Model(object, metaclass=ABCMeta):
             "use the grid property of the GridModel class instead.",
             DeprecationWarning,
         )
-        if len(self._staticmaps) == 0:
-            if self._read:
-                self.read_staticmaps()
+        if len(self._staticmaps) == 0 and self._read:
+            self.read_staticmaps()
         return self._staticmaps
 
     def set_staticmaps(
@@ -735,6 +744,7 @@ class Model(object, metaclass=ABCMeta):
         fn : str, optional
             filename relative to model root, by default "staticmaps/staticmaps.nc"
         """
+        self._assert_read_mode
         for ds in self._read_nc(fn, **kwargs).values():
             self.set_staticmaps(ds)
 
@@ -750,11 +760,13 @@ class Model(object, metaclass=ABCMeta):
         fn : str, optional
             filename relative to model root, by default 'staticmaps/staticmaps.nc'
         """
-        nc_dict = dict()
-        if len(self._staticmaps) > 0:
-            # _write_nc requires dict - use dummy key
-            nc_dict.update({"staticmaps": self._staticmaps})
-        self._write_nc(nc_dict, fn, **kwargs)
+        if len(self._staticmaps) == 0:
+            self.logger.debug("No staticmaps data found, skip writing.")
+        else:
+            self._assert_write_mode
+            # _write_nc requires dict - use dummy 'staticmaps' key
+            nc_dict = {"staticmaps": self._staticmaps}
+            self._write_nc(nc_dict, fn, **kwargs)
 
[…]
