### 2022-10-31 #4809 @elbeejay
:wave: @LejoFlores, @devincowan, and @rcabell I wanted to reach out and ask if you have any interest in reviewing this submission to JOSS titled: _"HydroMT: Automated and reproducible model building and analysis"_. You all have expertise at the intersection of hydrology and modeling, with an emphasis on modular code designs (as I understand it) via your involvement in the WRF-Hydro and HydroShare projects in particular. Please let me know if you'd be interested in serving as a reviewer, or if you know of other qualified folks who might be available.

At JOSS we do open checklist-driven reviews; peer-review criteria can be viewed [here](https://joss.readthedocs.io/en/latest/review_criteria.html). This issue is a "pre-review issue" which we use to find peer-reviewers. Once 3 reviewers are found, we will officially start the review in a dedicated GitHub issue. At present we are asking reviewers to complete reviews in 6 weeks, although this can be extended if needed.

If you are interested, please take a look at the journal's [conflict of interest policy](https://joss.readthedocs.io/en/latest/reviewer_guidelines.html?#joss-conflict-of-interest-policy) to ensure you do not have a conflict before agreeing to review this submission.

Thanks again for taking a moment to consider this, feel free to contact me if you have any questions about the JOSS review process. Please do not feel any pressure to accept this review request if you do not have the time or do not feel comfortable reviewing this software package, we appreciate and respect our peer-reviewers' time. If you cannot serve as a reviewer at this time but have a peer, collaborator, student, or colleague who might be available and would be a good fit for this submission, please let me know! This submission in particular seems appropriate to have a student conduct some or all of the peer-review process, as one of the target audiences for this package are students, as well as those with little to no Python knowledge beyond installation. So if interested we can figure out how to set up a "co-review" with a more junior or inexperienced member of the community.

Thanks,
Jay
