### 2022-11-08 Deltares/hydromt#243 @JannisHoch
JOSS: comments regarding paper
@DirkEilander @hboisgon @markhegnauer

Hi Dirk, Helene, Mark,
I have the pleasure to review the JOSS submission. While I still need to look into the code in detail, here already some comments on the actual JOSS software paper. Only a few things, it's overall pretty clear and well written.

**Summary**

- [ ] why can it be time-consuming to set-up hydro models?
- [ ] what are 'robust modelling practices'?
- [ ] I am struggling to understand how HydroMT is. Is it an interface that can be implemented into other models (i.e. models support HydroMT) Or can other models be added to HydroMT, i.e. HydroMT supports these models? Maybe it's just semantics, but it would be good to state it less ambigious.

**Statement of need**

Much of this section (esp. the first paragraph) covers the need for better data (e.g. through improved satellite instruments) and standardized input file-formats. In that way, HydroMT does not solve these challenges, but rather offers a solution to the syndroms not the problem. From that, I cannot directly derive the need for HydroMT.

The aspects mentioned in paragraphs 2 and 3 should thus be discussed more prominently. From what I understand, it is the easier set-up of models and workflows for a wide range of (different) models that makes HydroMT stands out. For what reasons would these aspects be needed? Maybe FAIR software principles and their increasingly important role in academia, or the need for model transparency in the policy-making context (just to name a few).

**Functionality**

In 'methods and workflow', a Figure 4.1 is mentioned which cannot be found in the paper.

While the Model API is model-agnostic, the ini-file is still model-specific, isn't it?
And would the user need to fill the ini-file with information or is this information added from the Data Adapter on-the-fly?
