### 2022-11-09 Deltares/hydromt#244 @JannisHoch
JOSS: HydroMT installation
As part of the JOSS review, I had to install HydroMT. Glad to report it worked as outlined in the documentation!

Two minor that may need to be improved:

1. The ``jupyter`` dependency is not installed by default in the HydroMT environment while it's needed to check the example notebooks. It's not a big thing to install it, but I would find it more elegant if it could be installed automatically.
2. It would be good to include the test whether the installation was succesful in the Installation Guide, i.e. the ``hydromt --models`` command.
