### 2022-11-09 9dd671b
fixed reading lists and none with config
