+++ hydromt/cli/cli_utils.py
@@ -7,6 +7,8 @@ import json
 import logging
 import click
 from ast import literal_eval
+from typing import Union, Dict
+from pathlib import Path
 
 from .. import config
 
@@ -69,16 +71,16 @@ def parse_json(ctx, param, value):
     return kwargs
 
 
-### general parsin methods ##
+### general parsing methods ##
 
 
-def parse_config(path=None, opt_cli=None):
+def parse_merge_config(path: Union[Path, str] = None, opt_cli: Dict = None) -> Dict:
+    """Parse config from ini `path` and combine with command line options `opt_cli`"""
     opt = {}
     if path is not None and isfile(path):
-        opt = config.configread(path, abs_path=True)
-        # make sure paths in config section are not abs paths
-        if "setup_config" in opt:
-            opt["setup_config"].update(config.configread(path).get("config", {}))
+        opt = config.configread(
+            path, abs_path=True, skip_eval_sections=["setup_config"]
+        )
     elif path is not None:
         raise IOError(f"Config not found at {path}")
     if opt_cli is not None:
