### 2022-11-17 Deltares/hydromt#244 @hboisgon
Hi Jannis, thanks for the feedback!
Tor un the example we recommend a different installation with the right yml containing jupyter notebook. But I see we did not include that file in the examples installation guide.
We prefer to leave the conda-install minimal to just the libraries you need to run hydromt.

Second point is also a good suggestion thanks!
