+++ paper.md
@@ -28,15 +28,17 @@ bibliography: paper.bib
 
 # Summary
 
-Many impact assessment studies rely on hydrological and hydrodynamic (hydro) models. The setup can be very time consuming and hard to reproduce if done manually. HydroMT (Hydro Model Tools) is an open-source Python package that aims to make the process of building models and analyzing model results automated and reproducible. The package provides a common interface to data and models, workflows to transform data into models based on (hydrological) GIS and statistical methods, and various methods to analyze model results. The user can describe a full model setup including its forcing in a single configuration file based on a sequence of workflows, making the process reproducible, fast, and modular. The package has been designed with an iterative, data-centered modeling process in mind. First-order models can be built for any location in the world by leveraging open global datasets. These models can later be improved by updating the input datasets with detailed local datasets. This iterative process enables the user to quickly get an initial model and analyze its result to then make informed decisions about the most relevant model improvements and/or required data collection and to kick-start discussions with stakeholders. Furthermore, model parameter maps or forcing data can easily be modified for model sensitivity analysis or model calibration to support robust modeling practices. HydroMT has successfully been implemented and applied for several model software packages through a plugin infrastructure and can easily be extended with new model software. Currently supported model software include the distributed rainfall-runoff model Wflow, the hydrodynamic flood model SFINCS, the water quality models D-Water Quality and D-Emissions and the flood impact model Delft-FIAT.
+Many impact assessment studies rely on hydrological and hydrodynamic (hydro) models. These models typically requires a large set of parameters that are derived from different datasets and hence the manual setup can be very time consuming and hard to reproduce. HydroMT (Hydro Model Tools) is an open-source Python package that aims to make the process of building models and analyzing model results automated and reproducible. The package provides a common interface to data and models, workflows to transform data into models based on (hydrological) GIS and statistical methods, and various methods to analyze model results. The user can describe a full model setup including its forcing in a single configuration file based on a sequence of workflows, making the process reproducible, fast, and modular. The package has been designed with an iterative, data-centered modeling process in mind. First-order models can be built for any location in the world by leveraging open global datasets. These models can later be improved by updating the input datasets with detailed local datasets. This iterative process enables the user to quickly get an initial model and analyze its result to then make informed decisions about the most relevant model improvements and/or required data collection and to kick-start discussions with stakeholders. Furthermore, model parameter maps or forcing data can easily be modified for model sensitivity analysis or model calibration to support these robust modeling practices. HydroMT is successfully being used for several model software packages through a plugin infrastructure that allow for model specific functionality such as readers and writers of model data formats and can easily be extended with new model software. Currently supported model software include the distributed rainfall-runoff model Wflow, the hydrodynamic flood model SFINCS, the water quality models D-Water Quality and D-Emissions and the flood impact model Delft-FIAT.
 
 # Statement of need
 
[…]
