+++ paper.bib
@@ -145,35 +145,6 @@
   language  = "en"
 }
 
-@ARTICLE{Van_Ormondt2020,
-  title    = "Delft Dashboard: a quick set-up tool for hydrodynamic models",
-  author   = "van Ormondt, Maarten and Nederhoff, Kees and van Dongeren, Ap",
-  abstract = "The open-source program Delft Dashboard (DDB) is a graphical user
-              interface designed to quickly create, edit input parameters and
-              visualize model inputs for a number of hydrodynamic models, using
-              private or publicly available local and global datasets. It
-              includes a number of toolboxes that facilitate the generation of
-              spatially varying inputs. These include new model schematizations
-              (grids, bathymetry, boundary conditions, etc.), cyclonic wind
-              fields and initial tsunami waves. The use of DDB can have
-              significant benefits. It can save modellers considerable time and
-              effort. Furthermore, the automated nature of both data collection
-              and pre-processing within the program reduces the likelihood of
-              errors that could occur when setting up models manually. Three
-              case studies are presented: simulation of tides in the North Sea,
-              storm surge and wave modelling under tropical cyclone conditions
-              and the simulation of a tsunami. The test cases show that models
-              created with DDB can be set up efficiently while maintaining a
-              predictive skill that is only slightly lower than that of
-              extensively calibrated models.",
-  journal  = "Journal of Hydroinformatics",
-  volume   =  22,
-  number   =  3,
-  pages    = "510--527",
-  month    =  feb,
-  year     =  2020
-}
-
 @ARTICLE{Bakker2016,
   title    = "Scripting {MODFLOW} Model Development Using Python and {FloPy}",
   author   = "Bakker, M and Post, V and Langevin, C D and Hughes, J D and
@@ -745,3 +716,109 @@
   institution = "Open Source Geospatial Foundation",
   year        =  2022
 }
+
+
+@MISC{Chue_Hong2021,
+  title     = "{FAIR} principles for Research Software ({FAIR4RS} principles)",
+  author    = "Chue Hong, Neil P and Katz, Daniel S and Barker, Michelle and
+               Lamprecht, Anna-Lena and Martinez, Carlos and Psomopoulos, Fotis
+               E and Harrow, Jen and Castro, Leyla Jael and Gruenpeter, Morane
+               and Martinez, Paula Andrea and Honeyman, Tom and Struck,
+               Alexander and Lee, Allen and Loewe, Axel and van Werkhoven, Ben
+               and Jones, Catherine and Garijo, Daniel and Plomp, Esther and
+               Genova, Francoise and Shanahan, Hugh and Leng, Joanna and
+               Hellstr{\"o}m, Maggie and Sandstr{\"o}m, Malin and Sinha,
+               Manodeep and Kuzak, Mateusz and Herterich, Patricia and Zhang,
+               Qian and Islam, Sharif and Sansone, Susanna-Assunta and Pollard,
+               Tom and Atmojo, Udayanto Dwi and Williams, Alan and Czerniak,
+               Andreas and Niehues, Anna and Fouilloux, Anne Claire and
+               Desinghu, Bala and Goble, Carole and Richard, C{\'e}line and
+               Gray, Charles and Erdmann, Chris and N{\"u}st, Daniel and
+               Tartarini, Daniele and Ranguelova, Elena and Anzt, Hartwig and
+               Todorov, Ilian and McNally, James and Moldon, Javier and
+               Burnett, Jessica and Garrido-S{\'a}nchez, Juli{\'a}n and
+               Belhajjame, Khalid and Sesink, Laurents and Hwang, Lorraine and
[…]
