+++ hydromt/data_adapter/dataframe.py
@@ -21,6 +21,7 @@ class DataFrameAdapter(DataAdapter):
         self,
         path,
         driver=None,
+        filesystem="local",
         nodata=None,
         rename={},
         unit_mult={},
@@ -45,6 +46,9 @@ class DataFrameAdapter(DataAdapter):
             :py:func:`~pandas.read_fwf`.
             By default the driver is inferred from the file extension and falls back to
             'csv' if unknown.
+        filesystem: {'local', 'gcs'}, optional
+            Filesystem where the data is stored (local, cloud, http etc.).
+            By default, local.
         nodata: (dictionary) float, int, optional
             Missing value number. Only used if the data has no native missing value.
             Multiple nodata values can be provided in a list and differentiated between
@@ -65,6 +69,7 @@ class DataFrameAdapter(DataAdapter):
         super().__init__(
             path=path,
             driver=driver,
+            filesystem=filesystem,
             nodata=nodata,
             rename=rename,
             unit_mult=unit_mult,
