+++ tests/test_forcing.py
@@ -1,0 +1,844 @@
+import pandas as pd
+import xarray as xr
+import numpy as np
+import re
+import logging
+from typing import Union
+
+from .. import _has_pyeto
+
+if _has_pyeto():
+    import pyeto
+
+logger = logging.getLogger(__name__)
+
+
+def precip(
+    precip,
+    da_like,
+    clim=None,
+    freq=None,
+    reproj_method="nearest_index",
+    resample_kwargs={},
+    logger=logger,
+):
+    """Lazy reprojection of precipitation to model grid and resampling of time dimension to frequency.
+
+    Parameters
+    ----------
+    precip: xarray.DataArray
+        DataArray of precipitation forcing [mm]
+    da_like: xarray.DataArray or Dataset
+        DataArray of the target resolution and projection.
+    clim: xarray.DataArray
+        DataArray of monthly precipitation climatology. If provided this is used to
+        to correct the precip downscaling.
+    freq: str, Timedelta
+        output frequency of time dimension
+    reproj_method: str, optional
+        Method for spatital reprojection of precip, by default 'nearest_index'
+    resample_kwargs:
+        Additional key-word arguments (e.g. label, closed) for time resampling method
+
+    Returns
+    --------
+    p_out: xarray.DataArray (lazy)
+        processed precipitation forcing
+    """
+    if precip.raster.dim0 != "time":
+        raise ValueError(f'First precip dim should be "time", not {precip.raster.dim0}')
+    # downscale precip (lazy); global min of zero
+    p_out = np.fmax(precip.raster.reproject_like(da_like, method=reproj_method), 0)
+    # correct precip based on high-res monthly climatology
+    if clim is not None:
+        # make sure first dim is month
+        clim = clim.rename({clim.raster.dim0: "month"})
+        if not clim["month"].size == 12:
+            raise ValueError("Precip climatology does not contain 12 months.")
+        # set missings to NaN
+        clim = clim.raster.mask_nodata()
+        # calculate downscaling multiplication factor
+        clim_coarse = clim.raster.reproject_like(
+            precip, method="average"
+        ).raster.reproject_like(da_like, method="average")
+        clim_fine = clim.raster.reproject_like(da_like, method="average")
+        p_mult = xr.where(clim_coarse > 0, clim_fine / clim_coarse, 1.0).fillna(1.0)
+        # multiply with monthly multiplication factor
+        p_out = p_out.groupby("time.month") * p_mult
+    # resample time
+    p_out.name = "precip"
+    p_out.attrs.update(unit="mm")
+    if freq is not None:
+        resample_kwargs.update(upsampling="bfill", downsampling="sum", logger=logger)
+        p_out = resample_time(p_out, freq, conserve_mass=True, **resample_kwargs)
+    return p_out
+
+
+# use dem_model (from staticmaps) and dem_forcing (meteo) in ini file
+def temp(
+    temp,
+    dem_model,
+    dem_forcing=None,
+    lapse_correction=True,
+    freq=None,
+    reproj_method="nearest_index",
+    lapse_rate=-0.0065,
+    resample_kwargs={},
+    logger=logger,
+):
+    """Lazy reprojection of temperature to model grid using lapse_rate for downscaling,
+    and resampling of time dimension to frequency.
+
+    Parameters
+    ----------
+    temp: xarray.DataArray
+        DataArray of temperature forcing [°C]
+    dem_model: xarray.DataArray
+        DataArray of the target resolution and projection, contains elevation
+        data
+    dem_forcing: xarray.DataArray
+        DataArray of elevation at forcing resolution. If provided this is used
+        with `dem_model` to correct the temperature downscaling using a lapse rate
+    lapse_correction : bool, optional
[…]
