### 2022-12-12 bad15d9
Added a test equal to script for penman
