### 2022-12-12 9941e8f
Added the test_precip for the penman branch and a placeholder for the test_pet
