### 2022-12-14 0b64208
Updated pm_fao56. Exact same performance as old pyeto variant. Fixed the 'pet' function.
