+++ hydromt/workflows/forcing.py
@@ -2,14 +2,14 @@ import pandas as pd
 import xarray as xr
 import xarray.core.resample
 import numpy as np
-import pyet
 import re
 import logging
 from typing import Union
 
-from .. import _has_pyeto
+from .._compat import HAS_PYET
 
-if _has_pyeto():
+if HAS_PYET:
+    import pyet
     import pyeto
 
 logger = logging.getLogger(__name__)
@@ -273,19 +273,19 @@ def wind(
 
 
 def pet(
-    ds,
-    temp,
-    dem_model,
-    method="debruin",
-    press_correction=False,
-    wind_correction=True,
-    wind_altitude=10,
-    reproj_method="nearest_index",
-    lapse_rate=-0.0065,
-    freq=None,
-    resample_kwargs={},
+    ds: xarray.Dataset,
+    temp: xarray.DataArray,
+    dem_model: xarray.DataArray,
+    method: str ="debruin",
+    press_correction: bool=False,
+    wind_correction: bool=True,
+    wind_altitude: float=10,
+    reproj_method: str="nearest_index",
+    lapse_rate: float=-0.0065,
+    freq: str=None,
+    resample_kwargs: dict={},
     logger=logger,
-):
+    ) -> xarray.DataArray:
 
     """Determines reference evapotranspiration (lazy reprojection on model grid and resampling of time dimension to frequency).
 
@@ -344,23 +344,18 @@ def pet(
             reproj_method=reproj_method,
         )
     else:
-        if "press_msl" in ds_out:
-            ds_out = ds_out.rename({"press_msl": "press"})
-        elif _has_pyeto():
+        if "press_msl" in ds:
+            ds_out["press"] = (
+                ds["press_msl"].raster.reproject_like(dem_model, method=reproj_method)
+                )
+        elif HAS_PYET:
             # calculate pressure from elevation [kPa]
-            ds_out["press"] = xr.apply_ufunc(
-                pyeto.atm_pressure,
-                dem_model.where(dem_model.mask),
-                dask="parallelized",
-                output_dtypes=[float],
-                vectorize=True,
-                keep_attrs=True,
-            )
+            ds_out["press"] = pyet.calc_press(dem_model)
             # convert to hPa to be consistent with press function calculation:
             ds_out["press"] = ds_out["press"] * 10
         else:
-            raise ValueError(
-                "If 'press' is supplied and 'press_correction' is used, the pyeto package must be installed."
+            raise ModuleNotFoundError(
+                "If 'press' is supplied and 'press_correction' is not used, the pyet package must be installed."
             )
 
     timestep = to_timedelta(ds).total_seconds()
@@ -378,49 +373,49 @@ def pet(
         )
     elif method == "makkink":
         pet_out = pet_makkink(temp, ds_out["press"], ds_out["kin"], timestep=timestep)
-    elif "penman-monteith" in method and _has_pyeto():
+    elif "penman-monteith" in method:
         logger.info("Calculating Penman-Monteith ref evaporation")
         # Add wind
         # compute wind from u and v components at 10m (for era5)
-        if ("wind_u" in ds.data_vars) & ("wind_v" in ds.data_vars):
+        if ("u10" in ds.data_vars) & ("v10" in ds.data_vars):
             ds_out["wind"] = wind(
                 da_model=dem_model,
-                wind_u=ds_out["wind_u"],
-                wind_v=ds_out["wind_v"],
+                wind_u=ds["u10"],
+                wind_v=ds["v10"],
                 altitude=wind_altitude,
[…]
