### 2022-12-14 37a387c
move pygeos to optional dependencies #228
