+++ hydromt/models/__init__.py
@@ -7,7 +7,7 @@ from .model_network import *
 from .model_plugins import *
 from .. import _compat
 
-if _compat.HAS_XUGRID:
+if _compat.HAS_XUGRID and _compat.HAS_PYGEOS:
     from .model_mesh import MeshModel
 
 # expose global MODELS object which discovers and loads
