+++ envs/hydromt-test.yml
@@ -20,7 +20,7 @@ dependencies:
   - packaging
   - pcraster # optional
   - pyflwdir>=0.5.4
-  - pygeos>=0.8
+  - pygeos>=0.8 # optional dependency for mesh
   - pyproj
   - pytest # tests
   - pytest-cov # tests
