+++ envs/hydromt-min.yml
@@ -19,7 +19,6 @@ dependencies:
   - packaging
   - pip  # install
   - pyflwdir>=0.5.4
-  - pygeos>=0.8
   - pyproj
   - python>=3.8
   - rasterio
