### 2022-12-14 b95b54e
fix some deprecation warnings
