+++ .github/workflows/tests.yml
@@ -17,7 +17,7 @@ jobs:
       fail-fast: false
       matrix:
         os: ["ubuntu-latest" ] #, "macos-latest", "windows-latest"]
-        python-version: ["3.9"] # fix tests to support older versions
+        python-version: ["3.9", "3.10"]
     
     steps:
       - uses: actions/checkout@v2
