### 2022-12-14 b480c31
make sure hydromt keeps working for older shapely versions
