+++ tests/test_vector_utils.py
@@ -1,8 +1,0 @@
-# #%%
-# from hydromt.vector_utils import *
-
-# # %%
-# l = LineString([(0, 0), (0, 1)])
-# # %%
-# print(npoints(l, 5))
-# # %%
