### 2022-12-14 2573fea
try fixing malloc error after test CI
