+++ tests/test_forcing.py
@@ -4,9 +4,8 @@ import pytest
 from hydromt.data_catalog import DataCatalog
 from hydromt.raster import full_from_transform
 from hydromt.workflows.forcing import precip, pet
-import pandas as pd
+import numpy as np
 import xarray as xr
-
 
 def test_precip():
     cat = DataCatalog()
@@ -40,17 +39,14 @@ def test_pet():
     cat = DataCatalog()
     et_data = cat.get_rasterdataset("era5_daily_zarr")
     et_data["d2m"] -= 273.15
-    dem = cat.get_rasterdataset("era5_orography")
+    dem = cat.get_rasterdataset("era5_orography").squeeze("time").drop("time")
 
     peto = pet(
-        et_data
-        w["temp"]
+        et_data,
+        et_data["temp"],
         dem,
         method="penman-monteith_tdew"
         )
 
     assert peto.raster.shape == dem.raster.shape
-
-
-
-
+    np.testing.assert_almost_equal(peto.mean(), 0.57746, decimal=4)
