### 2022-12-19 814c3f3
Added test for geodataset with placeholder function
