+++ tests/test_geodataset.py
@@ -1,0 +1,5 @@
+import pytest
+from hydromt.geodataset import GeoDataset
+
+def test_geo():
+    pass
