### 2022-12-23 #4897 @elbeejay
@DirkEilander it seems like @JannisHoch is pleased with the submission. I noticed that there were two issues opened:

https://github.com/Deltares/hydromt/issues/243

and

https://github.com/Deltares/hydromt/issues/244

Given the praise from @JannisHoch above, as well as the fact that they've now completed their reviewer checklist, it seems that they are comfortable with the package as it stands.

I have heard from @LejoFlores and it sounds like they intend to complete their review in the next week or so.

Looking at @mcflugen's checklist it looks like their review is nearing completion as well.

Given the amount of time that has elapsed between your submission and when we are reaching the end of this review process, and what appears to be a non-trivial amount of development work on `HydroMT`, I wanted to let you know that you are able to expand the list of authors if you desire. There will be an opportunity to update the version of the software associated with the JOSS publication at the end of this process as we require a tagged and archived copy of the software to be released to link to the JOSS paper. I just wanted to let you know that we are flexible in those regards, and that we are nearing the end of a somewhat long review process.
