### 2022-12-28 #4897 @elbeejay
@DirkEilander do you think you could revise the citations in the paper to address the missing DOIs as noted by the bot in the above comment?

@mcflugen I'm guessing that will address your "references" checkbox. Regarding the "state of the field" can you provide a comment or some context to give some insight into what you are looking for (or what you wanted to read but didn't)?
