+++ paper.bib
@@ -1,9 +1,10 @@
-@TECHREPORT{Undrr2022,
+@TECHREPORT{Undrr2022-yx,
   title       = "Global Assessment Report on Disaster Risk Reduction 2022: Our
                  World at Risk: Transforming Governance for a Resilient Future",
   author      = "{UNDRR}",
   institution = "United Nations Office for Disaster Risk Reduction",
-  year        =  2022
+  year        =  2022,
+  isbn        = "9789212320281"
 }
 
 % The entry below contains non-ASCII chars that could not be converted
@@ -14,24 +15,6 @@
   author    = "Hawker, Laurence and Uhe, Peter and Paulo, Luntadila and Sosa,
                Jeison and Savage, James and Sampson, Christopher and Neal,
                Jeffrey",
-  abstract  = "Elevation data are fundamental to many applications, especially
-               in geosciences. The latest global elevation data contains forest
-               and building artifacts that limit its usefulness for
-               applications that require precise terrain heights, in particular
-               flood simulation. Here, we use machine learning to remove
-               buildings and forests from the Copernicus Digital Elevation
-               Model to produce, for the first time, a global map of elevation
-               with buildings and forests removed at 1 arc second (∼30 m) grid
-               spacing. We train our correction algorithm on a unique set of
-               reference elevation data from 12 countries, covering a wide
-               range of climate zones and urban extents. Hence, this approach
-               has much wider applicability compared to previous DEMs trained
-               on data from a single country. Our method reduces mean absolute
-               vertical error in built-up areas from 1.61 to 1.12 m, and in
-               forests from 5.15 to 2.88 m. The new elevation map is more
-               accurate than existing global elevation maps and will strengthen
-               applications and models where high quality global terrain
-               information is required.",
   journal   = "Environ. Res. Lett.",
   publisher = "IOP Publishing",
   volume    =  17,
@@ -39,7 +22,9 @@
   pages     = "024016",
   month     =  feb,
   year      =  2022,
-  language  = "en"
+  language  = "en",
+  issn      = "1748-9326",
+  doi       = "10.1088/1748-9326/ac4d4f"
 }
 
 @ARTICLE{Neal2021,
@@ -47,43 +32,17 @@
                inundation models",
   author    = "Neal, Jeffrey and Hawker, Laurence and Savage, James and Durand,
                Michael and Bates, Paul and Sampson, Christopher",
-  abstract  = "Abstract Flood inundation modeling across large data sparse
-               areas has been increasing in recent years, driven by a desire to
-               provide hazard information for a wider range of locations. The
-               sophistication of these models has steadily advanced over the
-               past decade due to improvements in remote sensing and modeling
-               capability. There are now several global flood models (GFMs)
-               that seek to simulate water surface dynamics across all rivers
-               and floodplains regardless of data scarcity. However, flood
-               models in data sparse areas lack river bathymetry because this
-               cannot be observed remotely, meaning that a variety of methods
-               for approximating river bathymetry have been developed from
-               uniform flow or downstream hydraulic geometry theory. We argue
-               that bathymetry estimation in these models should follow
-               gradually varying flow theory to account for both uniform and
-               nonuniform flows. We demonstrate that existing methods for
-               bathymetry estimation in GFMs are only accurate for kinematic
-               water surface profiles and are unable to simulate unbiased water
-               surface profiles for reaches with diffusive or shallow water
-               wave properties. The use of gradually varied flow theory to
-               estimate bathymetry in a GFM reduced model error compared to a
-               target water surface profile by 66\% and eliminated bias due to
-               backwater effects. For a large-scale test case in Mozambique
[…]
