+++ paper.bib
@@ -1,4 +1,4 @@
-@TECHREPORT{Undrr2022-yx,
+@TECHREPORT{Undrr2022,
   title       = "Global Assessment Report on Disaster Risk Reduction 2022: Our
                  World at Risk: Transforming Governance for a Resilient Future",
   author      = "{UNDRR}",
