### 2023-01-02 #4897 @DirkEilander
@elbeejay Thanks for pushing the review process forward! It would indeed be nice to link the paper to the latest released version. I've updated the references in the HydroMT to include DOIs where available.
