### 2023-01-02 #4897 @mcflugen
> @mcflugen I'm guessing that will address your "references" checkbox. Regarding the "state of the field" can you provide a comment or some context to give some insight into what you are looking for (or what you wanted to read but didn't)?

@elbeejay Sorry for the delay. Now that the holidays have passed, I'll finish up my review. I'll open an issue over at HydroMT with some suggested changes but, overall, I think it's a great contribution and there shouldn't be much in the way of requested changes.
