+++ tests/test_geodataset.py
@@ -1,4 +1,5 @@
 import pytest
+from hydromt import DataCatalog
 from hydromt.geodataset import GeoDataset
 
 def test_geo():
