+++ envs/hydromt-dev.yml
@@ -49,6 +49,7 @@ dependencies:
   - xugrid>=0.1.5 # optional for mesh
   - zarr
   - pip:
+    - pyet
     - sphinx_design  # docs
     - sphinx_autosummary_accessors  # docs
 
