### 2023-01-03 #4897 @elbeejay
> > @mcflugen I'm guessing that will address your "references" checkbox. Regarding the "state of the field" can you provide a comment or some context to give some insight into what you are looking for (or what you wanted to read but didn't)?
>
> @elbeejay Sorry for the delay. Now that the holidays have passed, I'll finish up my review. I'll open an issue over at HydroMT with some suggested changes but, overall, I think it's a great contribution and there shouldn't be much in the way of requested changes.

No worries, thanks for the update and thanks for reviewing this submission @mcflugen.

@DirkEilander once @mcflugen is able to finish their review and open an issue, it sounds like you may just have a few minor things to edit. Once that is done we will move on with accepting and getting this submission published; I am going to be removing @LejoFlores from this review issue.
