+++ tests/test_forcing.py
@@ -7,6 +7,7 @@ from hydromt.workflows.forcing import precip, pet
 import numpy as np
 import xarray as xr
 
+
 def test_precip():
     cat = DataCatalog()
     p_precip = cat.get_rasterdataset("era5", variables="precip")  # era hourly
@@ -35,18 +36,14 @@ def test_precip():
     pout_freq = precip(p_precip, grid, freq="H")
     assert pout_freq.sizes["time"] == 313
 
+
 def test_pet():
     cat = DataCatalog()
     et_data = cat.get_rasterdataset("era5_daily_zarr")
     et_data["d2m"] -= 273.15
     dem = cat.get_rasterdataset("era5_orography").squeeze("time").drop("time")
 
-    peto = pet(
-        et_data,
-        et_data["temp"],
-        dem,
-        method="penman-monteith_tdew"
-        )
+    peto = pet(et_data, et_data["temp"], dem, method="penman-monteith_tdew")
 
     assert peto.raster.shape == dem.raster.shape
     np.testing.assert_almost_equal(peto.mean(), 0.57746, decimal=4)
