### 2023-01-03 Deltares/hydromt#260 @mcflugen
JOSS Review
@DirkEilander, @hboisgon, @markhegnauer Thank you for your JOSS submission for HydroMT. I think it's a great contribution and I don't see any major changes you need to make. I'll use this issue as a place to provide comments.

In the interest of getting things going, I have some minor edit suggestions to the paper that I'll provided now. The line numbers refer to those in the build of the paper [here](https://github.com/openjournals/joss-reviews/issues/4897#issuecomment-1360237107).

Minor edits:
* [x] Line 8: replace “requires” with “require”
* [x] Line 8: remove “that are”
* [x] Line 9: drop “the” before “manual setup”
* [x] Line 9: remove “very”
* [x] Line 31/32: remove “a lot of”
* [x] Line 35: I'm not sure "peculiar" is the correct word, maybe instead use “idiosyncratic”?
* [x] Line 60: replace “package” with “packages”
* [x] Line 63: “different models that”, not “difference models which”
* [x] Line 73: remove “globally”?
* [x] Line 79/80: replace “each of the HydroMT components” with “each HydroMT component”
* [x] Line 88/89: I think the sentence that begins with “tables for” should be reworded. Maybe “tables that can be used to, for example, convert land classes to roughness values”?
* [x] Line 100: “that” not “which”
* [x] Line 116: “Second, the user defines…”
* [x] Line 117: “that” not “which”
* [x] Line 119: “Third, the user configures…”
* [x] Line 124: “Finally, the user runs…”
* [x]  In *data_catalog.py*: I think the cross-reference is incorrect.
  -            See :py:func:`~hydromt.data_adapter.DataCatalog.from_yml` for accepted yaml format.
  +            See :py:func:`~hydromt.data_catalog.DataCatalog.from_yml` for accepted yaml format.
  ```
