### 2023-01-04 9707742
Fixed entrypoints check in test_api_model_components
