### 2023-01-04 522173b
Added zoom_level as keyword argument in 'get_rasterdataset'
