+++ hydromt/data_catalog.py
@@ -514,6 +514,7 @@ class DataCatalog(object):
         path_or_key: str,
         bbox: List = None,
         geom: gpd.GeoDataFrame = None,
+        zoom_level: Union[List, int] = 0,
         buffer: Union[float, int] = 0,
         align: bool = None,
         variables: Union[List, str] = None,
@@ -577,6 +578,7 @@ class DataCatalog(object):
             bbox=bbox,
             geom=geom,
             buffer=buffer,
+            zoom_level=zoom_level,
             align=align,
             variables=variables,
             time_tuple=time_tuple,
