### 2023-01-04 Deltares/hydromt#260 @mcflugen
A few additional comments:

* I think the paper could benefit from a definition of terms. In particular, what do the authors mean by a "model"? For me, the term "model" refers to what might be called a software model and is, I think, what the authors refer to in the Summary section. That is, a piece of software that ingests input data, solves a set of equations that represent a physical process, and then writes output data (other terms could be a computational model, numerical model or model kernel). Later on, though, I think the word "model" transitions more to the definition used in HydroMT's online docs. That is,
  > A set of files describing the schematization, forcing, states, simulation configuration and results for any supported model kernel and model classes.

  Instead of a definition of terms, perhaps using a qualifier with the word "model" that clarifies which meaning is being used would help.
* Thus far it seems as though HydroMT has been used with hydrological models. Is this a limitation of HydroMT? It certainly doesn't seem like it to me. It might be good to add a sentence somewhere that explicitly states the generality of HydroMT and that it can be used within any domain.
* The Summary section several times refers to HydroMT being used to analyze model results. In the body of the paper, however, there doesn't seem to be any explanation of how HydroMT can be used in this way. Since it sounds as though this it is one of the main capabilities of HydroMT a description of this capability would be useful.
* The Summary section states that HydroMT "provides a common interface to data and models". This is also exactly the goal of the Basic Model Interface[^1] (BMI). While the BMI defines an interface standard, the CSDMS Data Components [^2] provide implementations of the BMI for a collection of datasets. I think it would be interesting to hear how HydroMT compares to this other approach and how BMI could be used with HydroMT. Would it be possible, for example, to build a generalized HydroMT plugin that could access any BMI Data Component? Because this is my own work, I am somewhat uncomfortable making this suggestion so if this doesn't resonate with the authors or they find it off topic, I'm totally fine if they disregard this suggestion.

[^1]: https://doi.org/10.21105/joss.02317
[^2]: https://doi.org/10.5194/gmd-15-1413-2022
