### 2023-01-04 #4897 @mcflugen
@elbeejay I've made some comments in Deltares/hydromt#260 regarding the paper. Once those are addressed, I think the paper will be good to go.

I've tested the software on my Mac (intel) but not on any other platforms. Do you think it is necessary to test it on any additional platforms. If so, I can do that.
