### 2023-01-04 #4897 @elbeejay
@mcflugen thanks for doing that. I don't think we need additional testing on other platforms, looks like their tests run on ubuntu and I'd suspect at least 1 of the 14 contributors to the project is working on Windows.
