### 2023-01-06 45fa83c
'open_raster' now directly creates a dataarray
