+++ hydromt/data_adapter/rasterdataset.py
@@ -196,7 +196,9 @@ class RasterDatasetAdapter(DataAdapter):
             variables = np.atleast_1d(variables).tolist()
 
         kwargs = self.kwargs.copy()
-        fns = self.resolve_paths(time_tuple=time_tuple, zoom_level=zoom_level ,variables=variables)
+        fns = self.resolve_paths(
+            time_tuple=time_tuple, zoom_level=zoom_level, variables=variables
+        )
 
         # read using various readers
         if self.driver in ["netcdf"]:  # TODO complete list
