+++ paper.bib
@@ -398,3 +398,14 @@
   year      =  2012
 }
 
+@software{pandas2022,
+  author       = {The pandas development team},
+  title        = {pandas-dev/pandas: Pandas},
+  month        = nov,
+  year         = 2022,
+  note         = {If you use this software, please cite it as below.},
+  publisher    = {Zenodo},
+  version      = {v1.5.2},
+  doi          = {10.5281/zenodo.7344967},
+  url          = {https://doi.org/10.5281/zenodo.7344967}
+}
