### 2023-01-10 874f44c
'resolve_paths' fix: zoom_level is now internally a 'list'
