### 2023-01-10 f9404f9
Added function to create '.vrt' via GDAL
