### 2023-01-10 ab8c8fc
GDAL to '.vrt' is now seperate function in gis_utils
