### 2023-01-10 0c5b5d5
'resolve_paths' now accepts integer for zoom_level
