### 2023-01-11 #4897 @elbeejay
> @elbeejay maybe I'm missing something but it looks to me like they are still working on it as I don't see any changes to _paper.md_.

I agree - I'd just seen @DirkEilander checked off all of the individual items but looking at the `paper.md` file itself it appears the edits have yet to be pushed
