### 2023-01-11 2691d3f
fix pandas ref in bib
