+++ paper.bib
@@ -399,7 +399,7 @@
 }
 
 @software{pandas2022,
-  author       = {The pandas development team},
+  author       = "{The pandas development team}",
   title        = {pandas-dev/pandas: Pandas},
   month        = nov,
   year         = 2022,
