### 2023-01-11 Deltares/hydromt#260 @DirkEilander
Dear @mcflugen,

Many thanks for the useful review which has certainly improved the manuscript! Hereby a reply to your comments:

- We agree that we used the term model to refer both to the software and a model instance in the paper. We have tried to clarify this by adding a qualifier like "instance" of "software" to distinguish between both uses.
- While the package is not restricted to hydrological models, there are quite some hydrology-specific workflow available which makes the package most suitable for hydrological applications. We have clarified this in the paper by adding the following line to the end of the "statement of need" section:
  > While the package is sufficiently general to support model software outside the domain of hydro models, the focus of many workflows is on typical hydrological data and parameters and hence hydro models benefit most from the tool.
- The model analysis part was indeed not explained that well in the manuscript and has compared to the model building part also less functionality. We have added some explanation of this functionality, most importantly in the "Methods and workflows" component of the "functionality" section.
- I'm a big fan of BMI and use it mainly for online model coupling (e.g. with GLOFRIM https://github.com/openearth/glofrim) or interacting with model simulations. To my mind BMI and the HydroMT Model API serve different complementary purposes, where the power of HydroMT is in the pre- and postprocessing of model instances, the power of BMI is during the model simulation. However, the idea of a general HydroMT plugin for models with BMI is intriguing as it could allow users to view (and possibly interact with) a model instance on runtime or use the model software to read the model data. However, many model software store the data internally different from how the input data (e.g. gridded data is often flattened during the model initialization) which might make it difficult to make such a general plugin and still be able to the geo-context of the data.
