### 2023-01-12 e13a3a7
Added params in docstring of api_get_region (#207)
