### 2023-01-12 8322887
Apply suggestions from code review
Co-authored-by: DirkEilander <dirk.eilander@gmail.com>
