### 2023-01-12 b56ecff
Added to_osm for DataArray objects
