### 2023-01-12 1efe566
Added tests for 'to_xyz' and 'to_osm'
