+++ tests/test_raster.py
@@ -359,3 +359,38 @@ def test_zonal_stats():
         da.raster.zonal_stats(gdf, "asdf")
     with pytest.raises(IndexError, match="All geometries outside raster domain"):
         da.raster.zonal_stats(gdf.iloc[1:2], "mean")
+
+
+@pytest.fixture
+def dummy():
+    rast = raster.full_from_transform(
+        transform=[0.004166666666666666, 0.0, 0.0, 0.0, -0.004166666666666667, 0.0],
+        shape=(21600, 21600),
+        nodata=-9999,
+        name="dummy_tile_data",
+        crs=4326,
+    )
+    rast.values = np.random.randint(-5000, 5000, (21600, 21600))
+    return rast
+
+
+def test_to_xyz(tmpdir, dummy):
+    path = str(tmpdir)
+    dummy.raster.to_xyz(
+        f"{path}\\dummy_xyz",
+        px=1024,
+        zoomlevels=[0],
+    )
+    f = open(f"{path}\\dummy_xyz\\0\\filelist.txt", "r")
+    assert len(f.readlines()) == 484
+
+
+def test_to_osm(tmpdir, dummy):
+    path = str(tmpdir)
+    dummy.raster.to_osm(
+        f"{path}\\dummy_osm",
+        zl=4,
+        bbox=(0, -45, 45, 0),
+    )
+    f = open(f"{path}\\dummy_osm\\3\\filelist.txt", "r")
+    assert len(f.readlines()) == 4
