+++ hydromt/raster.py
@@ -8,9 +8,10 @@
 This module is an extension for xarray to provide rasterio capabilities
 to xarray datasets/dataarrays.
 """
-
+from __future__ import annotations
 import os
 from os.path import join, basename, dirname, isdir
+from typing import Any, Optional
 import numpy as np
 from shapely.geometry import box
 import pandas as pd
@@ -30,16 +31,18 @@ import tempfile
 import pyproj
 import logging
 import rioxarray
-
+import math
 from . import gis_utils, _compat
 
 logger = logging.getLogger(__name__)
-XDIMS = ("x", "longitude", "lon", "long")
-YDIMS = ("y", "latitude", "lat")
+XDIMS = ("xc", "x", "longitude", "lon", "long")
+YDIMS = ("yc", "y", "latitude", "lat")
 GEO_MAP_COORD = "spatial_ref"
 
 
-def full_like(other, nodata=None, lazy=False):
+def full_like(
+    other: xr.DataArray, nodata: float = None, lazy: bool = False
+) -> xr.DataArray:
     """Return a full object with the same grid and geospatial attributes as ``other``.
 
     Arguments
@@ -68,14 +71,24 @@ def full_like(other, nodata=None, lazy=False):
         attrs=other.attrs,
         crs=other.raster.crs,
         lazy=lazy,
+        shape=other.shape,
+        dims=other.dims,
     )
     da.raster.set_attrs(**other.raster.attrs)
     return da
 
 
 def full(
-    coords, nodata=np.nan, dtype=np.float32, name=None, attrs={}, crs=None, lazy=False
-):
+    coords,
+    nodata=np.nan,
+    dtype=np.float32,
+    name=None,
+    attrs={},
+    crs=None,
+    lazy=False,
+    shape=None,
+    dims=None,
+) -> xr.DataArray:
     """Return a full DataArray based on a geospatial coords dictionary.
 
     Arguments
@@ -95,6 +108,8 @@ def full(
         Coordinate Reference System. Accepts EPSG codes (int or str); proj (str or dict)
     lazy: bool, optional
         If True return DataArray with a dask rather than numpy array.
+    shape: tuple, optional
+        Length along (dim0, y, x) dimensions, of which the first is optional.
 
     Returns
     -------
@@ -102,8 +117,16 @@ def full(
         Filled DataArray
     """
     f = dask.array.empty if lazy else np.full
-    dims = tuple([d for d in coords])
-    shape = tuple([coords[dim].size for dim in dims])
+    if dims is None:
+        dims = tuple([d for d in coords])
+    if shape is None:
+        cs = next(iter(coords.values()))  # get first coordinate
+        if cs.ndim == 1:
+            shape = tuple([coords[dim].size for dim in dims])
+        else:  # rotated
+            shape = cs.shape
+            if hasattr(cs, "dims"):
+                dims = cs.dims
     data = f(shape, nodata, dtype=dtype)
     da = xr.DataArray(data, coords, dims, name, attrs)
     da.raster.set_nodata(nodata)
@@ -138,10 +161,11 @@ def full_from_transform(
     """
     if len(shape) not in [2, 3]:
         raise ValueError("Only 2D and 3D data arrays supported.")
-    _xcoords, _ycoords = gis_utils.affine_to_coords(transform, shape[-2:])
-    coords = {"y": _ycoords, "x": _xcoords}
+    coords = gis_utils.affine_to_coords(transform, shape[-2:], x_dim="x", y_dim="y")
+    dims = ("y", "x")
     if len(shape) == 3:
-        coords = {"dim0": np.arange(shape[0], dtype=int), **coords}
[…]
