+++ tests/test_gis_utils.py
@@ -5,6 +5,7 @@ import numpy as np
 from hydromt import gis_utils as gu
 from hydromt.raster import full_from_transform, RasterDataArray
 from rasterio.transform import from_origin
+from affine import Affine
 
 
 def test_crs():
@@ -23,18 +24,33 @@ def test_crs():
 def test_transform():
     transform = from_origin(0, 90, 1, 1)
     shape = (180, 360)
-    xs, ys = gu.affine_to_coords(transform, shape)
+    coords = gu.affine_to_coords(transform, shape)
+    xs, ys = coords["x"][1], coords["y"][1]
     assert np.all(ys == 90 - np.arange(0.5, shape[0]))
     assert np.all(xs == np.arange(0.5, shape[1]))
 
     # offset for geographic crs
     da = full_from_transform(transform, shape, crs=4326)
+    assert np.allclose(da.raster.origin, np.array([0, 90]))
     da1 = gu.meridian_offset(da, x_name="x")
     assert da1.raster.bounds[0] == -180
     da2 = gu.meridian_offset(da1, x_name="x", bbox=[170, 0, 190, 10])
-    assert da2.raster.bounds[0] == 170
+    assert da2.raster.bounds[0] == 0
     da3 = gu.meridian_offset(da1, x_name="x", bbox=[-190, 0, -170, 10])
-    assert da3.raster.bounds[2] == -170
+    assert da3.raster.bounds[2] == 0
+
+
+def test_transform_rotation():
+    # with rotation
+    transform = Affine.rotation(30) * Affine.scale(1, 2)
+    shape = (10, 5)
+    coords = gu.affine_to_coords(transform, shape)
+    xs, ys = coords["xc"][1], coords["yc"][1]
+    assert xs.ndim == 2 and ys.ndim == 2
+    da = full_from_transform(transform, shape, crs=4326)
+    assert da.raster.x_dim == "x"
+    assert da.raster.xcoords.ndim == 2
+    assert np.allclose(transform, da.raster.transform)
 
 
 def test_area_res():
