### 2023-01-16 eae0f96
Added 'from' and 'to' a ogr compliant Dataset/netCDF4 file
