+++ tests/test_geodataset.py
@@ -2,5 +2,6 @@ import pytest
 from hydromt import DataCatalog
 from hydromt.geodataset import GeoDataset
 
+
 def test_geo():
     pass
