+++ hydromt/geodataset.py
@@ -9,7 +9,9 @@ import geopandas as gpd
 import geopandas.array as geoarray
 from shapely.geometry import box
 from shapely.geometry.base import BaseGeometry
+from osgeo import osr
 import pyproj
+import shapely
 import logging
 
 from hydromt import gis_utils, raster
@@ -313,6 +315,117 @@ class GeoDataset(GeoBase):
         super(GeoDataset, self).__init__(xarray_obj)
 
     @staticmethod
+    def ogr_compliant(ds: xr.Dataset):
+        """Create a ogr compliant version of a xarray Dataset
+
+        Parameters
+        ----------
+        ds : xarray.Dataset
+            Dataset containing geospatial data
+
+        Returns
+        -------
+        xarray.Dataset
+            ogr compliant
+        """
+
+        _linkTable = {
+            int: "Integer64",
+            float: "Real",
+            str: "String",
+        }
+
+        def Type(val):
+            try:
+                s = eval(val)
+            except Exception:
+                s = val
+            return type(s)
+
+        def FieldType(lst):
+            if float or np.float64 in lst:
+                type = float
+            else:
+                type = int
+            if str in lst:
+                type = str
+            return type
+
+        wkt = [g.wkt for g in ds.geometry.values]
+
+        ## Determine Geometry type
+        from osgeo.ogr import CreateGeometryFromWkt
+
+        geom_types = [CreateGeometryFromWkt(g).GetGeometryName() for g in wkt]
+
+        if len(set(geom_types)) > 1:
+            i = ["MULTI" in g for g in geom_types].index(True)
+            geom_type = geom_types[i]
+        else:
+            geom_type = geom_types[0]
+
+        del geom_types
+
+        ## Create the geometry DataArray
+        ogc_wkt = xr.DataArray(
+            data=wkt,
+            dims="record",
+            attrs={
+                "long_name": "Geometry as ISO WKT",
+                "grid_mapping": "crs",
+            },
+        )
+
+        # Set spatial reference
+        srs = osr.SpatialReference()
+        srs.ImportFromEPSG(ds.geo.crs.to_epsg())
+
+        crs = xr.DataArray(
+            data=int(0),
+            attrs={
+                "long_name": "CRS definition",
+                "crs_wkt": srs.ExportToWkt(),
+                "spatial_ref": srs.ExportToWkt(),
+            },
+        )
+
+        out_ds = xr.Dataset({"ogc_wkt": ogc_wkt})
+
+        out_ds = out_ds.assign({"crs": crs})
+
+        for fld_header, da in ds.data_vars.items():
+            types = tuple(map(Type, da.values))
+
+            fld_type = FieldType(types)
+
+            temp_da = xr.DataArray(
+                data=da.values,
+                dims="record",
+                attrs={
+                    "ogr_field_name": f"{fld_header}",
+                    "ogr_field_type": _linkTable[fld_type],
+                },
+            )
+
+            if fld_type == str:
+                temp_da.attrs.update({"ogr_field_width": 100})
+            out_ds = out_ds.assign({f"{fld_header}": temp_da})
+
+            del temp_da
+
+        from osgeo import __version__ as GDAL_verion
+
+        out_ds = out_ds.assign_attrs(
+            {
+                "Conventions": "CF-1.6",
+                "GDAL": f"GDAL {GDAL_verion}",
+                "ogr_geometry_field": "ogc_wkt",
+                "ogr_layer_type": f"{geom_type}",
+            }
+        )
+        return out_ds
+
+    @staticmethod
     def from_gdf(gdf):
[…]
