### 2023-01-17 1b443dc
First draft geodataset unit test
