### 2023-01-17 4932bd6
Added test for ogr compliant geodataset
