### 2023-01-17 0ebad42
Added some extra checks to 'test_geodataset'
