### 2023-01-18 552c577
Removed default channel from 'tests.yml'
