+++ .github/workflows/tests.yml
@@ -25,7 +25,7 @@ jobs:
         with:
           python-version: ${{ matrix.python-version }}
           mamba-version: "*"
-          channels: conda-forge,defaults
+          channels: conda-forge
           channel-priority: true
           environment-file: envs/hydromt-test.yml
           activate-environment: hydromt-test
