### 2023-01-18 6297819
Revert "Removed default channel from 'tests.yml'"
This reverts commit 552c577ae1ba26c438929521d7c8bf1cc4a078b1.
