### 2023-01-18 #4897 @elbeejay
@DirkEilander and @mcflugen it looks like @DirkEilander has taken action to revise the paper and respond to the review comments raised by @mcflugen -- @mcflugen can you check, close https://github.com/Deltares/hydromt/issues/260, and complete your checklist if this is the case?
