### 2023-01-19 Deltares/hydromt#260 @mcflugen
@DirkEilander Sounds good. Thank you!

I found a couple more typos:
- [ ] LIne 19: "get an initial results" should be "get an initial result"
- [ ] Line 117: "how to build a model instances" should be "how to build a model instance"

Regarding BMI, I was thinking of datasets (i.e. ERA5, WAVEWATCHIII, etc.)—not software models—that expose a BMI through, for example, a Python class. Would it be possible to create a generalized HydroMT data adapter that can gather input data (typically, maybe exclusively, gridded datasets) for a software model from a BMI dataset? And, maybe more importantly, would it be useful? My thinking is that, with such an adapter, it may make it easier for you to add new datasets to your catalog if a BMI has already been implemented for those datasets. I'm really just thinking out loud here, no need to address this in your paper.
