### 2023-01-20 #4897 @elbeejay
@DirkEilander some minor comments/suggestions on the text:

Line 19 - "enables the user to quickly get an initial results", you'll want to revise this to be either "enables the user to quickly get initial results" or "enables the user to quickly get an initial result"

Line 25 - "can easily be extended with new model software" consider replacing "with" and using "to" instead

Line 31 - replace colon with comma in parenthetical (e.g., ) statement

Line 33 - add comma after "e.g.," in parenthetical statement

Line 51 - "Similarly, the LFPtools (Sosa et al., 2020) provides scripts", should be either, "Similarly, LFPtools (Sosa et al., 2020) provides scripts" or "Similarly, the LFPtools (Sosa et al., 2020) package provides scripts"

Line 52-53: "While these packages ..., the functionality" consider revising to "While these packages ..., functionality" or "While these packages ..., their functionality"

Line 63-64: Add comma parenthetical uses of "e.g." to make it "e.g.,"

Line 63 - Extra "n" in "Mannning"

Line 85 - Add comma to use of "e.g.,"

Line 92 - Capitalize "xarray" to be consistent with its capitalization on line 96

Line 94 - I believe "center lines" is a single word "centerlines" in this case

Line 98 - Consider pluralizing "forcing" to "forcings" to be consistent with pluralization of other items in list

Line 99 - Add comma after "i.e." to "i.e.,"

Line 106-107: Consider changing "readers and writers and workflows" to "readers, writers, and workflows"

Line 127 - Add comma after "i.e."

Line 127 - Is the text "setup methods" intentionally italicized?

I hope that is helpful, let me know if you have any questions about the suggested edits to the text.
