### 2023-01-23 Deltares/hydromt#260 @DirkEilander
Thanks @mcflugen for typos and suggestion!

I'm not too familiar with the BMI datasets, but if I understand it correctly it's basically a common API to retrieve datasets from different APIs/sources, am I correct? I do think that would be compatible with the HydroMT RasterDataset adapter and it becomes especially useful if different datasets can be accessed in the same way. Where can I find a full overview of the different datasets which are available? [This overview on the PyMT documentation](https://pymt.readthedocs.io/en/latest/models.html#data-components) seems to be outdated? I've opened a new issue to discuss this further: https://github.com/Deltares/hydromt/issues/271.
