### 2023-01-23 232c020
fix editorial suggestions in  https://github.com/openjournals/joss-reviews/issues/4897#issuecomment-1398452017
