### 2023-01-23 ec9f3bc
remove default channel; add  py3.10 for tests
