+++ .github/workflows/tests.yml
@@ -17,7 +17,7 @@ jobs:
       fail-fast: false
       matrix:
         os: ["ubuntu-latest" ] #, "macos-latest", "windows-latest"]
-        python-version: ["3.9"] # fix tests to support older versions
+        python-version: ["3.9", "3.10"] # fix tests to support older versions
     
     steps:
       - uses: actions/checkout@v2
@@ -25,7 +25,7 @@ jobs:
         with:
           python-version: ${{ matrix.python-version }}
           mamba-version: "*"
-          channels: conda-forge,defaults
+          channels: conda-forge
           channel-priority: true
           environment-file: envs/hydromt-test.yml
           activate-environment: hydromt-test
