### 2023-01-23 aba7a9b
try strict channel priority
