+++ .github/workflows/tests.yml
@@ -25,8 +25,8 @@ jobs:
         with:
           python-version: ${{ matrix.python-version }}
           mamba-version: "*"
-          channels: conda-forge
-          channel-priority: true
+          channels: conda-forge, defaults
+          channel-priority: strict
           environment-file: envs/hydromt-test.yml
           activate-environment: hydromt-test
 
