### 2023-01-23 1596640
try fixing mamba version to 1.1
