+++ .github/workflows/tests.yml
@@ -24,7 +24,7 @@ jobs:
       - uses: conda-incubator/setup-miniconda@v2
         with:
           python-version: ${{ matrix.python-version }}
-          mamba-version: "*"
+          mamba-version: "1.1"
           channels: conda-forge, defaults
           channel-priority: strict
           environment-file: envs/hydromt-test.yml
