### 2023-01-23 3e5071f
workaround for setup conda from https://github.com/conda-incubator/setup-miniconda/issues/274
