+++ .github/workflows/docs.yml
@@ -25,7 +25,7 @@ jobs:
         uses: conda-incubator/setup-miniconda@v2
         with:
           python-version: "3.10"
-          mamba-version: "1.1"
+          miniforge-variant: Mambaforge
           channels: conda-forge, defaults
           channel-priority: strict
           environment-file: envs/hydromt-dev.yml
