### 2023-01-23 468621e
Updated the geodataset to accept 'lonlat', 'wkt' and 'geometry'
