+++ hydromt/geodataset.py
@@ -15,14 +15,45 @@ import shapely
 import logging
 
 from hydromt import gis_utils, raster
+from hydromt.raster import XDIMS, YDIMS
+
+from osgeo import __version__ as GDAL_verion
 
 logger = logging.getLogger(__name__)
+
+def Type(val):
+    # try:
+    #     s = eval(val)
+    # except Exception:
+    #     s = val
+    return type(val)
+
+def FieldType(lst):
+    if float or np.float64 in lst:
+        type = float
+    else:
+        type = int
+    if str in lst:
+        type = str
+    return type
+
+_linkTable = {
+    int: "Integer64",
+    float: "Real",
+    str: "String",
+}
 
 #%%
 class GeoBase(raster.XGeoBase):
     def __init__(self, xarray_obj):
         super(GeoBase, self).__init__(xarray_obj)
-        self._geometry = None  # placeholder
+        self.set_meta()
+        self._geom_dims = {
+            "geometry": ("geometry",),
+            "wkt": ("ogc_wkt",),
+            "lonlat": (self.x_dim, self.y_dim),
+            None: None,
+        }
 
     @property
     def _all_names(self):
@@ -41,104 +72,23 @@ class GeoBase(raster.XGeoBase):
                 names.append(name)
         return names
 
-    def _discover_xy(self, x_name=None, y_name=None, index_dim=None):
-        """Set the spatial and index dimensions of the object.
-
-        Arguments
-        ----------
-        x_name, y_name, index_dim: str, optional
-            The name of the x, y and index dimensions.
-        """
-        # infer x dim
-        if x_name is None:
-            for name in raster.XDIMS:
-                if name in self._obj:
-                    dim0 = (
-                        index_dim if index_dim is not None else self._obj[name].dims[0]
-                    )
-                    if self._obj[name].ndim == 1 and self._obj[name].dims[0] == dim0:
-                        x_name = name
-                        index_dim = dim0
-                        break
-        # infer y dim
-        if y_name is None and x_name is not None:
-            for name in raster.YDIMS:
-                if name in self._obj:
-                    if self._obj[name].dims[0] == index_dim:
-                        y_name = name
-                        break
-        if (
-            x_name is not None
-            and y_name is not None
-            and self._obj[x_name].ndim == 1
-            and self._obj[x_name].dims == self._obj[y_name].dims
-        ):
-            self.set_attrs(x_name=x_name)
-            self.set_attrs(y_name=y_name)
-            self.set_attrs(index_dim=index_dim)
+    def set_meta(self):
+        self.x_dim = None
+        self.y_dim = None
+        self.index_dim = list(self._obj.dims)[0]
+        if any([item in XDIMS for item in self._all_names]):
+            self.dtype = "lonlat"
+            xind = [item in XDIMS for item in self._all_names].index(True)
+            self.x_dim = self._all_names[xind]
+            yind = [item in YDIMS for item in self._all_names].index(True)
+            self.y_dim = self._all_names[yind]
+            self.index_dim = list(self._obj[self.x_dim].dims)[0]
+        elif "ogc_wkt" in self._all_names:
+            self.dtype = "wkt" 
+        elif "geometry" in self._all_names:
+            self.dtype = "geometry"
         else:
-            self.attrs.pop("x_name", None)
-            self.attrs.pop("y_name", None)
-            self.attrs.pop("index_dim", None)
-
-    def _discover_geom(self, geom_name=None, index_dim=None):
-        # infer geom dim
-        if geom_name is None:
-            names = self._geom_names
-            if index_dim is not None:
-                names = [name for name in names if self._obj[name].dims[0] == index_dim]
[…]
