### 2023-01-23 #4897 @elbeejay
Thanks @DirkEilander - now we just need to sort out the metadata and versioning for the version of `HydroMT` that will be tied to this JOSS publication. Please let me know once you've been able to do the following:

- [ ] (if necessary) Re-release the package on PyPi/conda so that there is a pip-installable version of the package corresponding to what is going to be the JOSS-publication version

As we'd discussed above, this would be an acceptable time to alter the author list if needed. Let me know if you have any questions.

Thanks again to @mcflugen and @JannisHoch for reviewing this submission and providing constructive comments.
