### 2023-01-24 4dec37d
fix eval setup_config #274
