+++ tests/test_config.py
@@ -30,5 +30,13 @@ def test_config(tmpdir):
     assert cfdict["section1"] == cfdict1["section1"]
     assert isinstance(cfdict1["section2"]["path"], Path)
     assert isinstance(cfdict1["section2"]["path1"], str)
+    # by default paths in setup_config are not evaluated
     assert isinstance(cfdict1["setup_config"]["path"], str)
+    assert isinstance(cfdict1["setup_config"]["float"], float)
+    # return only str if skip_eval=True
+    cfdict1 = config.configread(config_fn, skip_eval=True)
+    for section in cfdict1:
+        assert all([isinstance(val, str) for val in cfdict1[section].values()])
+    # do not evaluate a specific section
+    cfdict1 = config.configread(config_fn, skip_eval_sections=["setup_config"])
     assert isinstance(cfdict1["setup_config"]["float"], str)
