+++ hydromt/config.py
@@ -22,7 +22,8 @@ def configread(
     noheader: bool = False,
     abs_path: bool = False,
     skip_eval: bool = False,
-    skip_eval_sections: List = ["setup_config"],
+    skip_eval_sections: List = [],
+    skip_abspath_sections: List = ["setup_config"],
 ) -> Dict:
     """Read configuration file and parse to (nested) dictionary.
     Values are evaluated and if possible parsed into python int, float, list or boolean types.
@@ -46,6 +47,9 @@ def configread(
         Skip evaluating argument values for python types, by default False
     skip_eval_sections: list, optional
         These sections are not evaluated for python types or absolute paths
+        if abs_path=True, by default []
+    skip_abspath_sections: list, optional
+        These sections are not evaluated for absolute paths
         if abs_path=True, by default ['update_config']
 
     Returns
@@ -78,7 +82,7 @@ def configread(
                 pass
             if isinstance(value, str) and len(value) == 0:
                 value = None
-            if abs_path:
+            if abs_path and section not in skip_abspath_sections:
                 if isinstance(value, str) and exists(join(root, value)):
                     value = Path(abspath(join(root, value)))
                 elif (
