### 2023-01-24 8300efa
geodataset: added reducer, to_xy, to_netcdf etc...
