+++ hydromt/geodataset.py
@@ -17,7 +17,7 @@ import logging
 from hydromt import gis_utils, raster
 from hydromt.raster import XDIMS, YDIMS
 
-from osgeo import __version__ as GDAL_verion
+from osgeo import __version__ as GDAL_VERSION
 
 logger = logging.getLogger(__name__)
 
@@ -37,13 +37,24 @@ def FieldType(lst):
         type = str
     return type
 
+def DeterGeomType(wkt):
+    from osgeo.ogr import CreateGeometryFromWkt
+    geom_types = [CreateGeometryFromWkt(g).GetGeometryName() for g in wkt]
+
+    if len(set(geom_types)) > 1:
+        i = ["MULTI" in g for g in geom_types].index(True)
+        geom_type = geom_types[i]
+    else:
+        geom_type = geom_types[0]
+    del geom_types
+    return geom_type
+
 _linkTable = {
     int: "Integer64",
     float: "Real",
     str: "String",
 }
 
-#%%
 class GeoBase(raster.XGeoBase):
     def __init__(self, xarray_obj):
         super(GeoBase, self).__init__(xarray_obj)
@@ -73,6 +84,15 @@ class GeoBase(raster.XGeoBase):
         return names
 
     def set_meta(self):
+        """Determines the meta of the GeoDataset/ GeoDataArray
+        Is invoked at Dataset/ DataArray creaton.
+
+        INCLUDES
+        --------
+        object.x_dim
+        object.y_dim
+        object.index_dim
+        """
         self.x_dim = None
         self.y_dim = None
         self.index_dim = list(self._obj.dims)[0]
@@ -92,10 +112,12 @@ class GeoBase(raster.XGeoBase):
 
     @property
     def index(self):
+        """Return the index values"""
         return self._obj[self.index_dim]
 
     @property
     def length(self):
+        """Return the length of the index array"""
         return self._obj[self.index_dim].size
 
     @property
@@ -104,7 +126,19 @@ class GeoBase(raster.XGeoBase):
         return self.geometry.total_bounds
 
     @property
-    def geometry(self):
+    def geometry(self) -> geoarray:
+        """_summary_
+
+        Returns
+        -------
+        geoarray
+            Array containing geometries
+
+        Raises
+        ------
+        ValueError
+            Unknown geometry format
+        """
         if self.dtype is None:
             raise ValueError("Unknown geometry format in Dataset")
         if self.dtype == "geometry":
@@ -131,17 +165,7 @@ class GeoBase(raster.XGeoBase):
         wkt = [g.wkt for g in self.geometry]
 
         ## Determine Geometry type
-        from osgeo.ogr import CreateGeometryFromWkt
-
-        geom_types = [CreateGeometryFromWkt(g).GetGeometryName() for g in wkt]
-
-        if len(set(geom_types)) > 1:
-            i = ["MULTI" in g for g in geom_types].index(True)
-            geom_type = geom_types[i]
-        else:
-            geom_type = geom_types[0]
-
-        del geom_types
+        geom_type = DeterGeomType(wkt)
 
         ## Create the geometry DataArray
         ogc_wkt = xr.DataArray(
@@ -179,7 +203,7 @@ class GeoBase(raster.XGeoBase):
 
         Returns
         -------
-        da: xarray.DataArray
+        xr.DataArray
             DataArray with transformed geospatial coordinates
         """
         if self.crs is None:
@@ -194,8 +218,42 @@ class GeoBase(raster.XGeoBase):
         obj.vector.set_crs(dst_crs)
         return obj
 
-    def to_lonlat(self):
-        pass
+    def to_xy(self) -> xr.Dataset:
+        """Converts Dataset/ DataArray to x,y structure
+
+        Returns
+        -------
+        xr.Dataset
+            x,y
+
+        Raises
+        ------
[…]
