+++ hydromt/raster.py
@@ -13,7 +13,7 @@ import os
 from os.path import join, basename, dirname, isdir
 from typing import Any, Optional
 import numpy as np
-from shapely.geometry import box
+from shapely.geometry import box, Polygon
 import pandas as pd
 import geopandas as gpd
 import xarray as xr
@@ -231,6 +231,13 @@ class XGeoBase(object):
                 if key.startswith(self.x_dim):
                     xcoords = self._obj.coords[key]
                     break
+        if xcoords.ndim == 2 and list(xcoords.dims).index(self.x_dim) != 1:
+            raise ValueError(
+                "Invalid raster: dimension order wrong. Fix using"
+                f'".transpose(..., {self.y_dim}, {self.x_dim})"'
+            )
+        if xcoords.size < 2 or (xcoords.ndim == 2 and xcoords.shape[1] < 2):
+            raise ValueError(f"Invalid raster: less than 2 cells in x_dim {self.x_dim}")
         return xcoords
 
     @property
@@ -242,6 +249,13 @@ class XGeoBase(object):
                 if key.startswith(self.y_dim):
                     ycoords = self._obj.coords[key]
                     break
+        if ycoords.ndim == 2 and list(ycoords.dims).index(self.y_dim) != 0:
+            raise ValueError(
+                "Invalid raster: dimension order wrong. Fix using"
+                f'".transpose(..., {self.y_dim}, {self.x_dim})"'
+            )
+        if ycoords.size < 2 or (ycoords.ndim == 2 and ycoords.shape[0] < 2):
+            raise ValueError(f"Invalid raster: less than 2 cells in y_dim {self.y_dim}")
         return ycoords
 
     def set_spatial_dims(self, x_dim=None, y_dim=None) -> None:
@@ -424,40 +438,50 @@ class XRasterBase(XGeoBase):
     def box(self) -> gpd.GeoDataFrame:
         """Return :py:meth:`~geopandas.GeoDataFrame` of bounding box"""
         crs = self.crs
-        if crs is None and crs.to_epsg() is not None:
+        if crs is not None and crs.to_epsg() is not None:
             crs = crs.to_epsg()  # not all CRS have an EPSG code
-        return gpd.GeoDataFrame(geometry=[box(*self.bounds)], crs=crs)
+        transform = self.transform
+        rs = np.array([0, self.height, self.height, 0, 0])
+        cs = np.array([0, 0, self.width, self.width, 0])
+        xs, ys = transform * (cs, rs)
+        return gpd.GeoDataFrame(geometry=[Polygon([*zip(xs, ys)])], crs=crs)
 
     @property
     def res(self) -> tuple[float, float]:
-        """Return resolution (x, y) tuple."""
+        """Return resolution (x, y) tuple.
+        NOTE: rotated rasters with a negative dx are not supported
+        """
         xs, ys = self.xcoords.data, self.ycoords.data
         dx, dy = 0, 0
         if xs.ndim == 1:
             dx = xs[1] - xs[0]
             dy = ys[1] - ys[0]
-            # assert np.allclose(np.diff(ys), dy)
-            # assert np.allclose(np.diff(xs), dx)
         elif xs.ndim == 2:
             ddx0 = xs[1, 0] - xs[0, 0]
             ddy0 = ys[1, 0] - ys[0, 0]
             ddx1 = xs[0, 1] - xs[0, 0]
             ddy1 = ys[0, 1] - ys[0, 0]
-            dy = np.sign(ys[1, 0] - ys[0, 0]) * math.hypot(ddx0, ddy0)
-            dx = np.sign(xs[0, 1] - xs[0, 0]) * math.hypot(ddx1, ddy1)
[…]
