### 2023-01-26 d6a1ed6
comment out topology group for forcing as not used at the moment
