### 2023-01-27 8986046
Updated example reading tabular data notebook and created dummy data in example/data folder together with a example data catalog (#217)
