+++ examples/data/tabular_data_catalog.yml
@@ -1,0 +1,9 @@
+example_csv_data:
+  path: example_csv_data.csv
+  data_type: DataFrame
+  driver: csv
+  
+dummy_excel_data:
+  path: dummy_excel_data.xlsx
+  data_type: DataFrame
+  driver: xlsx
