### 2023-01-27 2b79c2e
Removed pygeos entirely as a dependency
