+++ envs/hydromt-test.yml
@@ -20,7 +20,6 @@ dependencies:
   - packaging
   - pcraster # optional
   - pyflwdir>=0.5.4
-  - pygeos>=0.8 # optional dependency for mesh
   - pyproj
   - pytest # tests
   - pytest-cov # tests
@@ -29,7 +28,7 @@ dependencies:
   - requests
   - rioxarray
   - scipy
-  - shapely
+  - shapely>=2.0.0
   - xarray
   - xugrid>=0.1.5 # optional for mesh
   - zarr
