+++ envs/hydromt-dev.yml
@@ -30,7 +30,6 @@ dependencies:
   - pip
   - pydata-sphinx-theme  # docs
   - pyflwdir>=0.5.4
-  - pygeos>=0.8 # optional dependency for mesh
   - pyproj
   - pytest # tests
   - pytest-cov # tests
@@ -43,7 +42,7 @@ dependencies:
   - scipy
   - testpath  # tests
   - toml
-  - shapely
+  - shapely>=2.0.0
   - sphinx # docs
   - xarray
   - xugrid>=0.1.5 # optional for mesh
