### 2023-01-27 53416f5
restore set_spatial_dims; use GeoSeries for geometry property
