+++ hydromt/geodataset.py
@@ -1,17 +1,13 @@
-#!/usr/bin/env python
-# -*- coding: utf-8 -*-
 """"""
-#%%
+from __future__ import annotations
+from typing import Union
 import numpy as np
 import xarray as xr
-import pandas as pd
 import geopandas as gpd
-import geopandas.array as geoarray
-from shapely.geometry import box
+from geopandas import GeoSeries, GeoDataFrame
 from shapely.geometry.base import BaseGeometry
-from osgeo import osr
-import pyproj
 import shapely
+import pyproj
 import logging
 
 from hydromt import gis_utils, raster
@@ -21,12 +17,14 @@ from osgeo import __version__ as GDAL_VERSION
 
 logger = logging.getLogger(__name__)
 
+
 def Type(val):
     # try:
     #     s = eval(val)
     # except Exception:
     #     s = val
     return type(val)
+
 
 def FieldType(lst):
     if float or np.float64 in lst:
@@ -37,8 +35,10 @@ def FieldType(lst):
         type = str
     return type
 
+
 def DeterGeomType(wkt):
     from osgeo.ogr import CreateGeometryFromWkt
+
     geom_types = [CreateGeometryFromWkt(g).GetGeometryName() for g in wkt]
 
     if len(set(geom_types)) > 1:
@@ -49,22 +49,17 @@ def DeterGeomType(wkt):
     del geom_types
     return geom_type
 
+
 _linkTable = {
     int: "Integer64",
     float: "Real",
     str: "String",
 }
 
+
 class GeoBase(raster.XGeoBase):
     def __init__(self, xarray_obj):
         super(GeoBase, self).__init__(xarray_obj)
-        self.set_meta()
-        self._geom_dims = {
-            "geometry": ("geometry",),
-            "wkt": ("ogc_wkt",),
-            "lonlat": (self.x_dim, self.y_dim),
-            None: None,
-        }
 
     @property
     def _all_names(self):
@@ -73,42 +68,191 @@ class GeoBase(raster.XGeoBase):
             names = names + [n for n in self._obj.data_vars]
         return names
 
-    @property
-    def _geom_names(self):
-        names = []
-        for name in self._all_names:
+    def _get_geom_names_types(self, geom_name: str = None) -> tuple[list, list]:
+        names, types = [], []
+        dvars = self._all_names if geom_name is None else [geom_name]
+        for name in dvars:
             if self._obj[name].ndim == 1 and isinstance(
                 self._obj[name][0].values.item(), BaseGeometry
             ):
                 names.append(name)
-        return names
+                types.append("geom")
+            elif self._obj[name].ndim == 1 and isinstance(
+                self._obj[name][0].values.item(), str
+            ):
+                try:
+                    shapely.wkt.loads(self._obj[name][0].values.item())
+                    names.append(name)
+                    types.append("wkt")
+                except Exception:
+                    pass
+        return names, types
 
-    def set_meta(self):
-        """Determines the meta of the GeoDataset/ GeoDataArray
-        Is invoked at Dataset/ DataArray creaton.
-
-        INCLUDES
-        --------
-        object.x_dim
-        object.y_dim
-        object.index_dim
+    def _discover_xy(self, x_name=None, y_name=None, index_dim=None):
+        """Set the spatial and index dimensions of the object.
+        Arguments
+        ----------
+        x_name, y_name, index_dim: str, optional
+            The name of the x, y and index dimensions.
         """
-        self.x_dim = None
-        self.y_dim = None
-        self.index_dim = list(self._obj.dims)[0]
-        if any([item in XDIMS for item in self._all_names]):
-            self.dtype = "lonlat"
-            xind = [item in XDIMS for item in self._all_names].index(True)
-            self.x_dim = self._all_names[xind]
[…]
