+++ hydromt/cli/main.py
@@ -46,6 +46,16 @@ arg_baseroot = click.argument(
     type=click.Path(resolve_path=True, dir_okay=True, file_okay=False),
 )
 
+region_opt = click.option(
+    "-r",
+    "--region",
+    type=str,
+    default= "{}",
+    callback=cli_utils.parse_json,
+    help="Set the region in which to search for data via a json."
+    " e.g. {'subbasin': <path-to-file>, 'strord': 5}",
+)
+
 verbose_opt = click.option("--verbose", "-v", count=True, help="Increase verbosity.")
 
 quiet_opt = click.option("--quiet", "-q", count=True, help="Decrease verbosity.")
@@ -71,6 +81,14 @@ deltares_data_opt = click.option(
     is_flag=True,
     default=False,
     help='Shortcut to add the "deltares_data" catalog',
+)
+
+overwrite_opt = click.option(
+    "-o",
+    "--overwrite",
+    is_flag=True,
+    default=False,
+    help="Flag: whether or not to overwrite existing files",
 )
 
 ## MAIN
@@ -105,37 +123,25 @@ def main(ctx, models):  # , quiet, verbose):
     type=str,
 )
 @arg_root
-@click.argument(
-    "REGION",
-    type=str,
-    callback=cli_utils.parse_json,
-)
-@click.option(
-    "-r",
-    "--res",
-    type=float,
-    default=None,
-    help=f"Model resolution in model src.",
-)
-@click.option("--build-base/--build-all", default=False, help="Deprecated!")
 @opt_cli
 @opt_config
+@region_opt
 @data_opt
 @deltares_data_opt
-@quiet_opt
+@overwrite_opt
 @verbose_opt
+@quiet_opt
 @click.pass_context
 def build(
     ctx,
     model,
     model_root,
-    region,
-    res,
-    build_base,
     opt,
     config,
+    region,
     data,
     dd,
+    overwrite,
     verbose,
     quiet,
 ):
@@ -159,11 +165,6 @@ def build(
         "build", join(model_root, "hydromt.log"), log_level=log_level, append=False
     )
     logger.info(f"Building instance of {model} model at {model_root}.")
-    if build_base:
-        warnings.warn(
-            'The "build-base" flag has been deprecated, modify the ini file instead.',
-            DeprecationWarning,
-        )
     logger.info(f"User settings:")
     opt = cli_utils.parse_config(config, opt_cli=opt)
     kwargs = opt.pop("global", {})
@@ -182,7 +183,7 @@ def build(
             **kwargs,
         )
         # build model
-        mod.build(region, res, opt=opt)
+        mod.build(region, opt=opt)
     except Exception as e:
         logger.exception(e)  # catch and log errors
         raise
