### 2023-01-30 1c13514
Extra check for 'shapely 2.0' and environment var for 'pygeos'
