### 2023-01-30 53bb778
add example to user guide and update changelog
