+++ tests/test_model.py
@@ -300,7 +300,7 @@ def test_networkmodel(network_model, tmpdir):
         network_model.network
 
 
-@pytest.mark.skipif(not hydromt._compat.HAS_XUGRID, reason="Xugrid not installed.")
+@pytest.mark.skipif(hasattr(hydromt, "MeshModel"), reason="Xugrid not installed.")
 def test_meshmodel(mesh_model, tmpdir):
     MeshModel = MODELS.load("mesh_model")
     assert "mesh" in mesh_model.api
@@ -317,7 +317,7 @@ def test_meshmodel(mesh_model, tmpdir):
     assert equal, errors
 
 
-@pytest.mark.skipif(not hydromt._compat.HAS_XUGRID, reason="Xugrid not installed.")
+@pytest.mark.skipif(hasattr(hydromt, "MeshModel"), reason="Xugrid not installed.")
 def test_meshmodel_setup(griduda, world, tmpdir):
     MeshModel = MODELS.load("mesh_model")
     dc_param_fn = join(DATADIR, "parameters_data.yml")
