+++ hydromt/_compat.py
@@ -5,12 +5,20 @@ __all__ = []
 HAS_XUGRID = False
 HAS_PCRASTER = False
 HAS_SHAPELY20 = False
+HAS_PYGEOS = False
 
 try:
     from shapely import __version__ as SH_VERSION
 
     if Version(SH_VERSION) >= Version("2.0.0"):
         HAS_SHAPELY20 = True
+except ImportError:
+    pass
+
+try:
+    import pygeos
+
+    HAS_PYGEOS = True
 except ImportError:
     pass
 
