+++ hydromt/_compat.py
@@ -3,7 +3,7 @@ from packaging.version import Version
 __all__ = []
 
 HAS_XUGRID = False
-HAS_PCRASTER = False
+HAS_PCRASTER = True  # don't check PCRASTER compat for now, see below
 HAS_SHAPELY20 = False
 HAS_PYGEOS = False
 
@@ -22,12 +22,13 @@ try:
 except ImportError:
     pass
 
-try:
-    import pcraster
+# causes malloc / corrupted size errors on linux & github CI
+# try:
+#     import pcraster
 
-    HAS_PCRASTER = True
-except ImportError:
-    pass
+#     HAS_PCRASTER = True
+# except ImportError:
+#     pass
 
 
 try:
