+++ hydromt/geodataset.py
@@ -11,55 +11,14 @@ import pyproj
 import logging
 
 from hydromt import gis_utils, raster
-from hydromt.raster import XDIMS, YDIMS
-
-from osgeo import __version__ as GDAL_VERSION
 
 logger = logging.getLogger(__name__)
-
-
-def Type(val):
-    # try:
-    #     s = eval(val)
-    # except Exception:
-    #     s = val
-    return type(val)
-
-
-def FieldType(lst):
-    if float or np.float64 in lst:
-        type = float
-    else:
-        type = int
-    if str in lst:
-        type = str
-    return type
-
-
-def DeterGeomType(wkt):
-    from osgeo.ogr import CreateGeometryFromWkt
-
-    geom_types = [CreateGeometryFromWkt(g).GetGeometryName() for g in wkt]
-
-    if len(set(geom_types)) > 1:
-        i = ["MULTI" in g for g in geom_types].index(True)
-        geom_type = geom_types[i]
-    else:
-        geom_type = geom_types[0]
-    del geom_types
-    return geom_type
-
-
-_linkTable = {
-    int: "Integer64",
-    float: "Real",
-    str: "String",
-}
 
 
 class GeoBase(raster.XGeoBase):
     def __init__(self, xarray_obj):
         super(GeoBase, self).__init__(xarray_obj)
+        self._geometry = None
 
     @property
     def _all_names(self):
@@ -121,12 +80,12 @@ class GeoBase(raster.XGeoBase):
         ):
             self.set_attrs(x_name=x_name)
             self.set_attrs(y_name=y_name)
-            self.set_attrs(geom_type="xy")
+            self.set_attrs(geom_format="xy")
             self.set_attrs(index_dim=index_dim)
         else:
             self.attrs.pop("x_name", None)
             self.attrs.pop("y_name", None)
-            self.attrs.pop("geom_type", None)
+            self.attrs.pop("geom_format", None)
             self.attrs.pop("index_dim", None)
 
     def _discover_geom(self, geom_name=None, index_dim=None):
@@ -149,11 +108,11 @@ class GeoBase(raster.XGeoBase):
 
         if geom_name is not None:
             self.set_attrs(geom_name=geom_name)
-            self.set_attrs(geom_type=types[names.index(geom_name)])
+            self.set_attrs(geom_format=types[names.index(geom_name)])
             self.set_attrs(index_dim=self._obj[geom_name].dims[0])
         else:
             self.attrs.pop("geom_name", None)
-            self.attrs.pop("geom_type", None)
+            self.attrs.pop("geom_format", None)
             self.attrs.pop("index_dim", None)
 
     def set_spatial_dims(
@@ -162,69 +121,24 @@ class GeoBase(raster.XGeoBase):
         x_name=None,
         y_name=None,
         index_dim=None,
-        geom_type=None,
+        geom_format=None,
     ):
-        if geom_type != "xy":
+        if geom_format != "xy":
             self._discover_geom(geom_name=geom_name, index_dim=index_dim)
+            self.attrs.pop("x_name", None)
+            self.attrs.pop("y_name", None)
         if "geom_name" not in self.attrs:
             self._discover_xy(x_name=x_name, y_name=y_name, index_dim=index_dim)
+            self.attrs.pop("geom_name", None)
         elif "geom_name" not in self.attrs:
             raise ValueError("No geometry data found.")
 
-    def set_geometry(
-        self,
-        geometry: GeoSeries,
-        geom_type="geom",
-        geom_name="geometry",
-        x_name="x",
-        y_name="y",
-        index_dim=None,
-        drop_old_geometry=True,
-    ) -> Union[xr.DataArray, xr.Dataset]:
-        if index_dim is None:
[…]
