### 2023-02-02 82adfc4
Added 'error.py' for custom Exception classes
