+++ data/predefined_catalogs.yml
@@ -14,6 +14,7 @@ deltares_data:
 artifact_data:
    urlpath: https://github.com/DirkEilander/hydromt-artifacts/releases/download/{version}/data.tar.gz
    versions:
+      v0.0.8: v0.0.8
       v0.0.7: v0.0.7
       v0.0.6: v0.0.6
    notes: This data archive contains a sample dataset for the Piave basin (North Italy) to be used for tests and docs/demo purposes.
