+++ data/predefined_catalogs.yml
@@ -5,6 +5,7 @@
 #    versions:
 #       <calendar_version>: <git hash / tag>
 #    notes: <notes used in > 
+# if new versions of the data_catalogs are created, the current file predefined_catalogs.yml should be edited in main (not in the branch). 
 deltares_data:
    urlpath: https://raw.githubusercontent.com/Deltares/hydromt/{version}/data/catalogs/deltares_data.yml
    versions:
@@ -15,6 +16,7 @@ deltares_data:
 artifact_data:
    urlpath: https://github.com/DirkEilander/hydromt-artifacts/releases/download/{version}/data.tar.gz
    versions:
+      v0.0.8: v0.0.8
       v0.0.7: v0.0.7
       v0.0.6: v0.0.6
    notes: This data archive contains a sample dataset for the Piave basin (North Italy) to be used for tests and docs/demo purposes.
