### 2023-02-03 e69d9cf
cli: removed '--res', made region optional '-r'
