### 2023-02-03 95dae40
Updated docstring/cli string of 'build'
