+++ pyproject.toml
@@ -15,7 +15,6 @@ dependencies = [
     "dask",
     "entrypoints",
     "geopandas>=0.10",
-    "gcsfs",
     "gdal>=3.1",
     "numpy",
     "netcdf4",
