+++ envs/hydromt-min.yml
@@ -11,7 +11,6 @@ dependencies:
   - dask
   - entrypoints
   - flit>=3.4  # install
-  - gcsfs
   - gdal>=3.1
   - geopandas>=0.10
   - netcdf4
