### 2023-02-07 fa7fd15
Readded ogr_compliant plus small stuff
