+++ hydromt/vector.py
@@ -1,28 +1,26 @@
-#!/usr/bin/env python
-# -*- coding: utf-8 -*-
-"""Convenience methods for reading and writing point  objects.
-"""
-
+""""""
+from __future__ import annotations
+from typing import Union
 import numpy as np
-import geopandas as gpd
 import xarray as xr
-from shapely.geometry import box
+import geopandas as gpd
+from geopandas import GeoSeries, GeoDataFrame
+from shapely.geometry.base import BaseGeometry
+import shapely
 import pyproj
 import logging
 
-from . import gis_utils, raster
+from hydromt import gis_utils, raster
+
+from osgeo import __version__ as GDAL_VERSION
 
 logger = logging.getLogger(__name__)
 
 
 class GeoBase(raster.XGeoBase):
-    """This is a vector extension for xarray.DataArray 1D time series data
-    with geographical location information"""
-
     def __init__(self, xarray_obj):
         super(GeoBase, self).__init__(xarray_obj)
-        # placeholder
-        self._sindex = None
+        self._geometry = None
 
     @property
     def _all_names(self):
@@ -31,90 +29,159 @@ class GeoBase(raster.XGeoBase):
             names = names + [n for n in self._obj.data_vars]
         return names
 
-    @property
-    def x_dim(self):
-        # NOTE overwrites raster.XGeoBase.x_dim
-        if self.get_attrs("x_dim") not in self._all_names:
-            self.set_spatial_dims()
-        return self.attrs["x_dim"]
+    def _get_geom_names_types(self, geom_name: str = None) -> tuple[list, list]:
+        names, types = [], []
+        dvars = self._all_names if geom_name is None else [geom_name]
+        for name in dvars:
+            if self._obj[name].ndim == 1 and isinstance(
+                self._obj[name][0].values.item(), BaseGeometry
+            ):
+                names.append(name)
+                types.append("geom")
+            elif self._obj[name].ndim == 1 and isinstance(
+                self._obj[name][0].values.item(), str
+            ):
+                try:
+                    shapely.wkt.loads(self._obj[name][0].values.item())
+                    names.append(name)
+                    types.append("wkt")
+                except Exception:
+                    pass
+        return names, types
 
-    @property
-    def y_dim(self):
-        # NOTE overwrites raster.XGeoBase.y_dim
-        if self.get_attrs("y_dim") not in self._all_names:
-            self.set_spatial_dims()
-        return self.attrs["y_dim"]
-
-    def set_spatial_dims(self, x_dim=None, y_dim=None, index_dim=None):
+    def _discover_xy(self, x_name=None, y_name=None, index_dim=None):
         """Set the spatial and index dimensions of the object.
-
         Arguments
         ----------
-        x_dim, y_dim, index_dim: str, optional
+        x_name, y_name, index_dim: str, optional
             The name of the x, y and index dimensions.
         """
-        # NOTE overwrites raster.XGeoBase.set_spatial_dims
-        # infer x and index dims
-        _obj = self._obj
-        names = list(_obj.coords.keys())
-        if isinstance(_obj, xr.Dataset):
-            names = names + list(_obj.data_vars.keys())
-        if x_dim is None:
-            for dim in raster.XDIMS:
-                if dim in names:
-                    dim0 = index_dim if index_dim is not None else _obj[dim].dims[0]
-                    if _obj[dim0].dims[0] == dim0:
-                        x_dim = dim
+        # infer x dim
+        if x_name is None:
+            for name in raster.XDIMS:
+                if name in self._obj:
+                    dim0 = (
+                        index_dim if index_dim is not None else self._obj[name].dims[0]
+                    )
+                    if self._obj[name].ndim == 1 and self._obj[name].dims[0] == dim0:
+                        x_name = name
[…]
