+++ hydromt/geodataset.py
@@ -1,658 +1,0 @@
-""""""
-from __future__ import annotations
-from typing import Union
-import numpy as np
-import xarray as xr
-import geopandas as gpd
-from geopandas import GeoSeries, GeoDataFrame
-from shapely.geometry.base import BaseGeometry
-import shapely
-import pyproj
-import logging
-
-from hydromt import gis_utils, raster
-
-from osgeo import __version__ as GDAL_VERSION
-
-logger = logging.getLogger(__name__)
-
-
-class GeoBase(raster.XGeoBase):
-    def __init__(self, xarray_obj):
-        super(GeoBase, self).__init__(xarray_obj)
-        self._geometry = None
-
-    @property
-    def _all_names(self):
-        names = [n for n in self._obj.coords]
-        if isinstance(self._obj, xr.Dataset):
-            names = names + [n for n in self._obj.data_vars]
-        return names
-
-    def _get_geom_names_types(self, geom_name: str = None) -> tuple[list, list]:
-        names, types = [], []
-        dvars = self._all_names if geom_name is None else [geom_name]
-        for name in dvars:
-            if self._obj[name].ndim == 1 and isinstance(
-                self._obj[name][0].values.item(), BaseGeometry
-            ):
-                names.append(name)
-                types.append("geom")
-            elif self._obj[name].ndim == 1 and isinstance(
-                self._obj[name][0].values.item(), str
-            ):
-                try:
-                    shapely.wkt.loads(self._obj[name][0].values.item())
-                    names.append(name)
-                    types.append("wkt")
-                except Exception:
-                    pass
-        return names, types
-
-    def _discover_xy(self, x_name=None, y_name=None, index_dim=None):
-        """Set the spatial and index dimensions of the object.
-        Arguments
-        ----------
-        x_name, y_name, index_dim: str, optional
-            The name of the x, y and index dimensions.
-        """
-        # infer x dim
-        if x_name is None:
-            for name in raster.XDIMS:
-                if name in self._obj:
-                    dim0 = (
-                        index_dim if index_dim is not None else self._obj[name].dims[0]
-                    )
-                    if self._obj[name].ndim == 1 and self._obj[name].dims[0] == dim0:
-                        x_name = name
-                        index_dim = dim0
-                        break
-        # infer y dim
-        if y_name is None and x_name is not None:
-            for name in raster.YDIMS:
-                if name in self._obj:
-                    if self._obj[name].dims[0] == index_dim:
-                        y_name = name
-                        break
-        if (
-            x_name is not None
-            and y_name is not None
-            and self._obj[x_name].ndim == 1
-            and self._obj[x_name].dims == self._obj[y_name].dims
-        ):
-            self.set_attrs(x_name=x_name)
-            self.set_attrs(y_name=y_name)
-            self.set_attrs(geom_format="xy")
-            self.set_attrs(index_dim=index_dim)
-        else:
-            self.attrs.pop("x_name", None)
-            self.attrs.pop("y_name", None)
-            self.attrs.pop("geom_format", None)
-            self.attrs.pop("index_dim", None)
-
-    def _discover_geom(self, geom_name=None, index_dim=None):
-        # check /infer geom dim
-        names, types = self._get_geom_names_types(geom_name=geom_name)
-        if geom_name is None:
-            if index_dim is not None:
-                names = [name for name in names if self._obj[name].dims[0] == index_dim]
-            # priority for geometry named variable/coord
-            if "geometry" in names:
-                geom_name = "geometry"
[…]
