+++ hydromt/vector.py
@@ -500,20 +500,28 @@ class GeoBase(raster.XGeoBase):
         gdf: geopandas.GeoDataFrame
             GeoDataFrame
         """
-        gdf = self.geometry.to_frame("geometry")
-        sdims = [self.y_name, self.x_name, self.index_dim, self.geom_name]
-        for name in self._all_names:
-            dims = self._obj[name].dims
+        if isinstance(self._obj, xr.DataArray):
+            if self._obj.name is None:
+                self._obj.name = "data"
+            obj = self._obj.to_dataset()
+        else:
+            obj = self._obj
+        gdf = obj.vector.geometry.to_frame("geometry")
+        sdims = [obj.vector.y_name, obj.vector.x_name, obj.vector.index_dim, obj.vector.geom_name]
+        for name in obj.vector._all_names:
+            dims = obj[name].dims
             if name not in sdims:
                 # keep 1D variables with matching index_dim
-                if len(dims) == 1 and dims[0] == self.index_dim:
-                    gdf[name] = self._obj[name].values
+                if len(dims) == 1 and dims[0] == obj.vector.index_dim:
+                    gdf[name] = obj[name].values
                 # keep reduced data variables
-                elif reducer is not None and self.index_dim in self._obj[name].dims:
+                elif (reducer is not None
+                    and obj.vector.index_dim in obj[name].dims):
                     rdims = [
-                        dim for dim in self._obj[name].dims if dim != self.index_dim
+                        dim for dim in obj[name].dims if dim != obj.vector.index_dim
                     ]
-                    gdf[name] = self._obj[name].reduce(reducer, dim=rdims)
+                    gdf[name] = obj[name].reduce(reducer, dim=rdims)
+        del obj
         return gdf
 
     def to_netcdf(
@@ -536,7 +544,9 @@ class GeoBase(raster.XGeoBase):
         if ogr_compliant:
             self.ogr_compliant(reducer=reducer).to_netcdf(path, engine="netcdf4", **kwargs)
         else:
-            self._obj.to_netcdf(path, engine="netcdf4", **kwargs)
+            obj = self.update_geometry(geom_format="wkt", geom_name="ogc_wkt")
+            obj.to_netcdf(path, engine="netcdf4", **kwargs)
+            del obj
 
 @xr.register_dataarray_accessor("vector")
 class GeoDataArray(GeoBase):
@@ -546,17 +556,50 @@ class GeoDataArray(GeoBase):
     # Constructers
     # i.e. from other datatypes or files
     @staticmethod
-    def from_gdf(gdf: GeoSeries | GeoDataFrame) -> xr.Dataset:
-        if isinstance(gdf, GeoSeries):
-            gdf = gdf.to_frame("geometry")
-        geom_name = gdf._geometry_column_name
-        index_dim = gdf.index.name
+    def from_gdf(
+        gdf: gpd.GeoDataFrame,
+        data: np.array = None,
+        name: str = None,
+        coords: dict = None,
+        dims: tuple = None,
+        index_dim: str = None,
+        keep_cols: bool = True,
+    ) -> xr.DataArray:
         if index_dim is None:
-            gdf.index.name = "index"
-        ds = gdf.to_xarray().set_coords(geom_name)
-        ds.vector.set_spatial_dims(geom_name=geom_name, index_dim=index_dim)
-        ds.vector.set_crs(gdf.crs)
-        return ds
+            index_dim = gdf.index.name if gdf.index.name is not None else "index"
+        geom_name = gdf.geometry.name
+        if dims is None:
+            dims = (index_dim, )
+        if len(dims) <= data.shape.__len__():
+            dims += tuple([
[…]
