### 2023-02-08 0eec603
fix small bug with zoom_level=None and tidy example for user guide
