### 2023-02-08 eebc197
fix dropping y_dim
