+++ tests/test_cli.py
@@ -39,18 +39,23 @@ def test_cli(tmpdir):
     )
 
     root = str(tmpdir.join("grid_model_region"))
-    r = CliRunner().invoke(
-        hydromt_cli,
-        [
-            "build",
-            "grid_model",
-            root,
-            "-r",
-            "{'bbox': [12.05,45.30,12.85,45.65]}",
-            "-vv",
-        ],
-    )
+    cmd = [
+        "build",
+        "grid_model",
+        root,
+        "-r",
+        "{'bbox': [12.05,45.30,12.85,45.65]}",
+        "-vv",
+    ]
+    r = CliRunner().invoke(hydromt_cli, cmd)
     assert os.path.isfile(os.path.join(root, "geoms", "region.geojson"))
+
+    # test force overwrite
+    with pytest.raises(IOError, match="Model dir already exists"):
+        r = CliRunner().invoke(hydromt_cli, cmd)
+        raise r.exception
+    r = CliRunner().invoke(hydromt_cli, cmd + ["--fo"])
+    assert r.exit_code == 0
 
     root = str(tmpdir.join("empty_region"))
     r = CliRunner().invoke(
