### 2023-02-09 b925791
'from_gdf' keyword args now actually do something
