+++ hydromt/vector.py
@@ -321,8 +321,7 @@ class GeoBase(raster.XGeoBase):
 
     # Internal conversion and selection methods
     # i.e. produces xarray.Dataset/ xarray.DataArray
-    def ogr_compliant(self, reducer = None):
-
+    def ogr_compliant(self, reducer=None):
         obj = self.update_geometry(geom_format="wkt", geom_name="ogc_wkt")
         obj["ogc_wkt"].attrs = {
             "long_name": "Geometry as ISO WKT",
@@ -394,7 +393,9 @@ class GeoBase(raster.XGeoBase):
         return self.update_geometry(geom_format="xy", x_name=x_name, y_name=y_name)
 
     def to_wkt(
-        self, ogr_compliant=False, reducer=None,
+        self,
+        ogr_compliant=False,
+        reducer=None,
     ) -> Union[xr.DataArray, xr.Dataset]:
         """Converts geometries in Dataset/DataArray to wkt strings.
 
@@ -482,7 +483,6 @@ class GeoBase(raster.XGeoBase):
     # Constructers
     # i.e. from other datatypes or files
 
-
     ## Output methods
     ## Either writes to files or other data types
     def to_gdf(self, reducer=None):
@@ -507,7 +507,12 @@ class GeoBase(raster.XGeoBase):
         else:
             obj = self._obj
         gdf = obj.vector.geometry.to_frame("geometry")
-        sdims = [obj.vector.y_name, obj.vector.x_name, obj.vector.index_dim, obj.vector.geom_name]
+        sdims = [
+            obj.vector.y_name,
+            obj.vector.x_name,
+            obj.vector.index_dim,
+            obj.vector.geom_name,
+        ]
         for name in obj.vector._all_names:
             dims = obj[name].dims
             if name not in sdims:
@@ -515,8 +520,7 @@ class GeoBase(raster.XGeoBase):
                 if len(dims) == 1 and dims[0] == obj.vector.index_dim:
                     gdf[name] = obj[name].values
                 # keep reduced data variables
-                elif (reducer is not None
-                    and obj.vector.index_dim in obj[name].dims):
+                elif reducer is not None and obj.vector.index_dim in obj[name].dims:
                     rdims = [
                         dim for dim in obj[name].dims if dim != obj.vector.index_dim
                     ]
@@ -542,11 +546,14 @@ class GeoBase(raster.XGeoBase):
             This makes it readable as a vector file in e.g. QGIS
         """
         if ogr_compliant:
-            self.ogr_compliant(reducer=reducer).to_netcdf(path, engine="netcdf4", **kwargs)
+            self.ogr_compliant(reducer=reducer).to_netcdf(
+                path, engine="netcdf4", **kwargs
+            )
         else:
             obj = self.update_geometry(geom_format="wkt", geom_name="ogc_wkt")
             obj.to_netcdf(path, engine="netcdf4", **kwargs)
             del obj
+
 
 @xr.register_dataarray_accessor("vector")
 class GeoDataArray(GeoBase):
@@ -569,18 +576,18 @@ class GeoDataArray(GeoBase):
             index_dim = gdf.index.name if gdf.index.name is not None else "index"
         geom_name = gdf.geometry.name
         if dims is None:
-            dims = (index_dim, )
+            dims = (index_dim,)
         if dims.__len__() <= data.shape.__len__():
-            dims += tuple([
-                f"dim{num}" for num in range(len(data.shape)-len(dims))
-            ])
+            dims += tuple([f"dim{num}" for num in range(len(data.shape) - len(dims))])
         elif dims.__len__() >= data.shape.__len__():
-            raise OverflowError(f"No of dimensions should not exceed those of\
-                 the data -> {dims.__len__()} vs {data.shape.__len__()}")
+            raise OverflowError(
[…]
