### 2023-02-09 851b67e
updated some docstrings
