### 2023-02-09 98a1358
fix path check for linux
