+++ hydromt/workflows/forcing.py
@@ -214,7 +214,6 @@ def pet(
     resample_kwargs={},
     logger=logger,
 ):
-
     """Determines reference evapotranspiration (lazy reprojection on model grid and resampling of time dimension to frequency).
 
     Parameters
@@ -310,7 +309,6 @@ def press_correction(
 
 
 def temp_correction(dem, lapse_rate=-0.0065):
-
     """Temperature correction based on elevation data.
 
     Parameters
