+++ hydromt/raster.py
@@ -1659,7 +1659,7 @@ class RasterDataArray(XRasterBase):
         self,
         root: str,
         px: int,
-        zoomlevels: list = [],
+        zoomlevels: list,
         fmt: str = "tif",
     ):
         """Export rasterdataset to tiles in a xyz structure
@@ -1670,9 +1670,14 @@ class RasterDataArray(XRasterBase):
             Path where the database will be saved
             Database yml will be put one directory above
         px : int
-            Amount of pixels per tile
-        zoomlevels : list, optional
+            Amount of pixels per tile in one direction
+            This determines the shape of the tiles
+            E.g. px = 256 -> shape = (256, 256)
+        zoomlevels : list
             Zoom levels to be put in the database
+        fmt : str, optional
+            File format
+            Either 'nc' or 'tif'
 
         Raises
         ------
@@ -1695,11 +1700,14 @@ class RasterDataArray(XRasterBase):
         pxs = [px * 2**num for num in range(nzl)]
 
         # Clean up zoomlevels list
+        if not zoomlevels:
+            raise ValueError("At least one zoomlevel is required")
         zoomlevels = [x for x in zoomlevels if x <= len(pxs) - 1]
+        zoomlevels.sort()
 
-        def tile_window(px):
+        def tile_window(shape, px):
             # Basic stuff
-            nr, nc = self._obj.shape
+            nr, nc = shape
             lu = product(range(0, nc, px), range(0, nr, px))
             ## create the window
             for l, u in lu:
@@ -1707,50 +1715,63 @@ class RasterDataArray(XRasterBase):
                 w = min(px, nc - l)
                 yield (l, u, w, h)
 
+        vrt = None
+        prev = 0
+        obj = self._obj.copy()
         for zl in zoomlevels:
-            sd = f"{root}\\{zl}"
+            diff = zl - prev
+            pxzl = px*(2**(diff))
+            if vrt is not None:
+                obj = xr.open_dataarray(vrt,engine="rasterio").squeeze("band")
+                obj = obj.drop("band")
+            sd = join(root,f"{zl}")
             folder(sd)
             # Write the raster paths to a text file
-            file = open(f"{sd}\\filelist.txt", "w")
+            file = open(join(sd,"filelist.txt"), "w")
 
-            for l, u, w, h in tile_window(pxs[zl]):
-                col = int(np.ceil(l / pxs[zl]))
-                row = int(np.ceil(u / pxs[zl]))
-                ssd = "{}\\{}".format(sd, col)
+            for l, u, w, h in tile_window(obj.shape, pxzl):
+                col = int(np.ceil(l / pxzl))
+                row = int(np.ceil(u / pxzl))
+                ssd = join(sd,f"{col}")
 
                 folder(ssd)
 
                 # create temp tile
-                temp = self._obj[u : u + h, l : l + w].load()
+                temp = obj[u : u + h, l : l + w]
                 # temp = self._obj.isel({self.y_dim:slice(u,u+h), self.x_dim: slice(l,l+w)})
                 if zl != 0:
-                    dst_transform = temp.raster.transform * temp.raster.transform.scale(
-                        2**zl
-                    )
-                    temp = temp.raster.reproject(
-                        dst_transform=dst_transform,
-                        dst_width=int(w / (2**zl)),
-                        dst_height=int(h / (2**zl)),
-                        method="average",
-                        # dst_crs=self.crs,
-                    )
[…]
