### 2023-02-09 c64287d
Updated documentation and example
