+++ docs/changelog.rst
@@ -13,7 +13,6 @@ Added
 -----
 - New get_region method in cli/api.py that returns a geojson representation of the parsed region `PR #209 <https://github.com/Deltares/hydromt/pull/209>`_
 - write raster (DataArray) to tiles in xyz stucture
-- write raster (DataArray) to tiles according to osm
 - add zoom_level to named arguments of path in the data catalog, to keep a single data source for the same data at different resolutions
 - add function in gis_utils to write '.vrt' using GDAL
 - fix bug related to opening named raster files #200
