+++ docs/api.rst
@@ -440,7 +440,6 @@ Raster writing methods
    :toctree: _generated
 
    DataArray.raster.to_xyz
-   DataArray.raster.to_osm
    DataArray.raster.to_raster
    Dataset.raster.to_mapstack
 
@@ -686,6 +685,7 @@ Raster
 .. autosummary::
    :toctree: _generated
 
+   gis_utils.create_vrt
    gis_utils.spread2d
    gis_utils.reggrid_area
    gis_utils.cellarea
