+++ docs/user_guide/intro.rst
@@ -115,6 +115,7 @@ Region                          Argument of the *build* and *clip* CLI methods t
    data_conventions.rst
    ../_examples/prep_data_catalog.ipynb
    ../_examples/export_data.ipynb
+   ../_examples/tiling_raster_data.ipynb
    ../_examples/reading_raster_data.ipynb
    ../_examples/reading_vector_data.ipynb
    ../_examples/reading_point_data.ipynb
