### 2023-02-09 2010e3c
small changes to to_gdf; update changelog; add docstrings
