+++ hydromt/vector.py
@@ -1,5 +1,5 @@
 from __future__ import annotations
-from typing import Union
+from typing import Union, Any
 import numpy as np
 import xarray as xr
 import geopandas as gpd
@@ -11,6 +11,7 @@ import logging
 
 from hydromt import gis_utils, raster
 
+# TODO try getting this version without importing osgeo to avoid conflicts with geopandas/rasterio
 from osgeo import __version__ as GDAL_VERSION
 
 logger = logging.getLogger(__name__)
@@ -23,12 +24,15 @@ class GeoBase(raster.XGeoBase):
 
     @property
     def _all_names(self):
+        """Return names of all variables and coordinates in the dataset/array"""
+        # TODO: move to geobase
         names = [n for n in self._obj.coords]
         if isinstance(self._obj, xr.Dataset):
             names = names + [n for n in self._obj.data_vars]
         return names
 
     def _get_geom_names_types(self, geom_name: str = None) -> tuple[list, list]:
+        """Discover coordinates with wkt/geom type data the dataset/array"""
         names, types = [], []
         dvars = self._all_names if geom_name is None else [geom_name]
         for name in dvars:
@@ -49,12 +53,7 @@ class GeoBase(raster.XGeoBase):
         return names, types
 
     def _discover_xy(self, x_name=None, y_name=None, index_dim=None):
-        """Set the spatial and index dimensions of the object.
-        Arguments
-        ----------
-        x_name, y_name, index_dim: str, optional
-            The name of the x, y and index dimensions.
-        """
+        """Discover xy type geometries in the dataset/array"""
         # infer x dim
         if x_name is None:
             for name in raster.XDIMS:
@@ -90,6 +89,7 @@ class GeoBase(raster.XGeoBase):
             self.attrs.pop("index_dim", None)
 
     def _discover_geom(self, geom_name=None, index_dim=None):
+        """Discover geom/wkt type geometries in the dataset/array"""
         # check /infer geom dim
         names, types = self._get_geom_names_types(geom_name=geom_name)
         if geom_name is None:
@@ -118,12 +118,23 @@ class GeoBase(raster.XGeoBase):
 
     def set_spatial_dims(
         self,
-        geom_name=None,
-        x_name=None,
-        y_name=None,
-        index_dim=None,
-        geom_format=None,
-    ):
+        geom_name: str = None,
+        x_name: str = None,
+        y_name: str = None,
+        index_dim: str = None,
+        geom_format: str = None,
+    ) -> None:
+        """Set the spatial and index dimensions of the object.
+
+        Arguments
+        ----------
+        x_name, y_name, geom_name: str, optional
+            The name of the x, y and geometry coordinate.
+        index_dim: str optional
+            The name of the geometry index dimension
+        geom_format: {'xy', 'wkt', 'geom'}
+            Geometry format
+        """
         if geom_format != "xy":
             self._discover_geom(geom_name=geom_name, index_dim=index_dim)
             self.attrs.pop("x_name", None)
@@ -131,25 +142,27 @@ class GeoBase(raster.XGeoBase):
         if "geom_name" not in self.attrs:
             self._discover_xy(x_name=x_name, y_name=y_name, index_dim=index_dim)
             self.attrs.pop("geom_name", None)
-        elif "geom_name" not in self.attrs:
+        if "geom_format" not in self.attrs:
             raise ValueError("No geometry data found.")
 
     @property
-    def geom_format(self):
+    def geom_format(self) -> str:
+        """Name of geometry coordinate; only for 'wkt' and 'geom' formats."""
         if self.get_attrs("geom_format") not in self._obj.dims:
             self.set_spatial_dims()
         if "geom_format" in self.attrs:
             return self.attrs["geom_format"]
 
     @property
-    def geom_name(self):
+    def geom_name(self) -> str:
[…]
