### 2023-02-09 385a116
Neatly the conversions within the function
