+++ hydromt/data_adapter/geodataset.py
@@ -217,7 +217,7 @@ class GeoDatasetAdapter(DataAdapter):
         ds_out = ds_out.rename(rm)
         # check spatial dims and make sure all are set as coordinates
         try:
-            # ds_out.vector.set_spatial_dims()
+            ds_out.vector.set_spatial_dims()
             idim = ds_out.vector.index_dim
             if idim not in ds_out:  # set coordinates for index dimension if missing
                 ds_out[idim] = xr.IndexVariable(idim, np.arange(ds_out.dims[idim]))
@@ -243,10 +243,7 @@ class GeoDatasetAdapter(DataAdapter):
         if geom is not None:
             bbox = geom.to_crs(4326).total_bounds
         if ds_out.vector.crs.to_epsg() == 4326:
-            w, e = (
-                ds_out.vector.geometry.total_bounds[0],
-                ds_out.vector.geometry.total_bounds[2],
-            )
+            e = ds_out.vector.geometry.total_bounds[2]
             if e > 180 or (bbox is not None and (bbox[0] < -180 or bbox[2] > 180)):
                 ds_out = gis_utils.meridian_offset(ds_out, ds_out.vector.x_name, bbox)
         if geom is not None:
