### 2023-02-09 3c51ba2
Updated documentation with the correct CLI calls and added note on new region flag on model_build page (#278)
