### 2023-02-09 ab0cbd0
small updates to_xyz incl test and example
