+++ hydromt/raster.py
@@ -1656,11 +1656,7 @@ class RasterDataArray(XRasterBase):
         return interp_array
 
     def to_xyz(
-        self,
-        root: str,
-        px: int,
-        zoomlevels: list,
-        fmt: str = "tif",
+        self, root: str, tile_size: int, zoomlevels: list, driver="GTiff", **kwargs
     ):
         """Export rasterdataset to tiles in a xyz structure
 
@@ -1669,44 +1665,25 @@ class RasterDataArray(XRasterBase):
         root : str
             Path where the database will be saved
             Database yml will be put one directory above
-        px : int
+        tile_size : int
             Amount of pixels per tile in one direction
             This determines the shape of the tiles
-            E.g. px = 256 -> shape = (256, 256)
+            E.g. tile_size = 256 -> shape = (256, 256)
         zoomlevels : list
             Zoom levels to be put in the database
-        fmt : str, optional
-            File format
-            Either 'nc' or 'tif'
-
-        Raises
-        ------
-        ValueError
-            Amount of pixels can not be larger than the original dataset
+        driver : str, optional
+            GDAL driver (e.g., 'GTiff' for geotif files), or 'netcdf4' for netcdf files.
+        **kwargs
+            Key-word arguments to write raster files
         """
-
         mName = os.path.normpath(os.path.basename(root))
 
-        def folder(path):
+        def create_folder(path):
             if not os.path.exists(path):
                 os.makedirs(path)
 
-        folder(root)
-
-        mdim = min(self._obj.shape)
-        if px > mdim:
-            raise ValueError("")
-        nzl = int(np.ceil((np.log10(mdim / px) / np.log10(2))))
-        pxs = [px * 2**num for num in range(nzl)]
-
-        # Clean up zoomlevels list
-        if not zoomlevels:
-            raise ValueError("At least one zoomlevel is required")
-        zoomlevels = [x for x in zoomlevels if x <= len(pxs) - 1]
-        zoomlevels.sort()
-
         def tile_window(shape, px):
-            # Basic stuff
+            """Yield (left, upper, width, height)"""
             nr, nc = shape
             lu = product(range(0, nc, px), range(0, nr, px))
             ## create the window
@@ -1717,50 +1694,56 @@ class RasterDataArray(XRasterBase):
 
         vrt = None
         prev = 0
+        nodata = self.nodata
         obj = self._obj.copy()
         for zl in zoomlevels:
             diff = zl - prev
-            pxzl = px*(2**(diff))
+            pxzl = tile_size * (2 ** (diff))
+
+            # read data from previous zoomlevel
             if vrt is not None:
-                obj = xr.open_dataarray(vrt,engine="rasterio").squeeze("band")
+                obj = xr.open_dataarray(vrt, engine="rasterio").squeeze("band")
                 obj = obj.drop("band")
-            sd = join(root,f"{zl}")
-            folder(sd)
+            x_dim, y_dim = obj.raster.x_dim, obj.raster.y_dim
+            obj = obj.chunk({x_dim: pxzl, y_dim: pxzl})
+
+            if pxzl > min(obj.shape):
+                logger.warning(
+                    f"Tiles at zoomlevel {zl} smaller than tile_size {tile_size}"
+                )
+
             # Write the raster paths to a text file
-            file = open(join(sd,"filelist.txt"), "w")
+            sd = join(root, f"{zl}")
+            create_folder(sd)
+            file = open(join(sd, "filelist.txt"), "w")
 
             for l, u, w, h in tile_window(obj.shape, pxzl):
                 col = int(np.ceil(l / pxzl))
[…]
