+++ hydromt/data_adapter/rasterdataset.py
@@ -1,6 +1,7 @@
 # -*- coding: utf-8 -*-
+import os
 from os.path import join
-import fsspec
+from fsspec.implementations import local
 import numpy as np
 import pandas as pd
 import xarray as xr
@@ -200,38 +201,59 @@ class RasterDatasetAdapter(DataAdapter):
         if variables:
             variables = np.atleast_1d(variables).tolist()
 
+        # Extract storage_options from kwargs to instantiate fsspec object correctly
+        if "storage_options" in self.kwargs:
+            kwargs = self.kwargs["storage_options"]
+            # For s3, anonymous connection still requires --no-sign-request profile to read the data
+            # setting environment variable works
+            if "anon" in kwargs:
+                os.environ["AWS_NO_SIGN_REQUEST"] = "YES"
+            else:
+                os.environ["AWS_NO_SIGN_REQUEST"] = "NO"
+        else:
+            kwargs = dict()
+        fns = self.resolve_paths(time_tuple=time_tuple, variables=variables, **kwargs)
+
         kwargs = self.kwargs.copy()
-        fns = self.resolve_paths(time_tuple=time_tuple, variables=variables)
+        # zarr can use storage options directly, the rest should be converted to file-like objects
+        if "storage_options" in kwargs and self.driver == "raster":
+            storage_options = kwargs.pop("storage_options")
+            fs = self.get_filesystem(**storage_options)
+            fns = [fs.open(f) for f in fns]
 
         # read using various readers
         if self.driver in ["netcdf"]:  # TODO complete list
-            if "preprocess" in kwargs:
-                preprocess = PREPROCESSORS.get(kwargs["preprocess"], None)
-                kwargs.update(preprocess=preprocess)
-            ds_out = xr.open_mfdataset(fns, decode_coords="all", **kwargs)
+            if self.filesystem == "local":
+                if "preprocess" in kwargs:
+                    preprocess = PREPROCESSORS.get(kwargs["preprocess"], None)
+                    kwargs.update(preprocess=preprocess)
+                ds_out = xr.open_mfdataset(fns, decode_coords="all", **kwargs)
+            else:
+                raise NotImplementedError(
+                    "Remote (cloud) RasterDataset not supported with driver netcdf."
+                )
         elif self.driver == "zarr":
-            # if len(fns) > 1:
-            #    raise ValueError(
-            #        "RasterDataset: Opening multiple zarr data files is not supported."
-            #    )
             if "preprocess" in kwargs:  # for zarr preprocess is done after reading
                 preprocess = PREPROCESSORS.get(kwargs.pop("preprocess"), None)
                 do_preprocess = True
             ds_lst = []
             for fn in fns:
-                # if "mapper" in kwargs:
-                #    mapper_type = kwargs.pop("mapper")
-                #    if mapper_type == "fsspec":
-                #        mapper = fsspec.get_mapper(str(fn))
                 ds = xr.open_zarr(fn, **kwargs)
                 if do_preprocess:
                     ds = preprocess(ds)
                 ds_lst.append(ds)
             ds_out = xr.merge(ds_lst)
         elif self.driver == "raster_tindex":
-            if np.issubdtype(type(self.nodata), np.number):
-                kwargs.update(nodata=self.nodata)
-            ds_out = io.open_raster_from_tindex(fns[0], bbox=bbox, geom=geom, **kwargs)
+            if self.filesystem == "local":
+                if np.issubdtype(type(self.nodata), np.number):
+                    kwargs.update(nodata=self.nodata)
+                ds_out = io.open_raster_from_tindex(
+                    fns[0], bbox=bbox, geom=geom, **kwargs
+                )
+            else:
+                raise NotImplementedError(
[…]
