### 2023-02-10 e8a494f
Add s3fs to test dependency
