+++ envs/hydromt-test.yml
@@ -30,8 +30,9 @@ dependencies:
   - requests
   - rioxarray
   - scipy
-  - universal_pathlib
   - shapely>=2.0.0
+  - s3fs
+  - universal_pathlib
   - xarray
   - xugrid>=0.1.5 # optional for mesh
   - zarr
