+++ data/catalogs/deltares_data.yml
@@ -1,6 +1,6 @@
 meta:
   root: p:/wflow_global/hydromt
-  version: 2022.8
+  version: 2023.2
 
 basin_atlas_level12_v10: 
   crs: 4326
