+++ docs/api.rst
@@ -597,8 +597,12 @@ High level methods
 
    DataArray.vector.from_gdf
    DataArray.vector.to_gdf
+   DataArray.vector.from_netcdf
+   DataArray.vector.to_netcdf
    Dataset.vector.from_gdf
-
+   Dataset.vector.to_gdf
+   Dataset.vector.from_netcdf
+   Dataset.vector.to_netcdf
 
 Attributes
 ==========
@@ -611,15 +615,48 @@ Attributes
    DataArray.vector.crs
    DataArray.vector.index_dim
    DataArray.vector.time_dim
-   DataArray.vector.y_dim
-   DataArray.vector.x_dim
-   DataArray.vector.xcoords
-   DataArray.vector.ycoords
+   DataArray.vector.x_name
+   DataArray.vector.y_name
+   DataArray.vector.geom_name
+   DataArray.vector.geom_type
+   DataArray.vector.geom_format
    DataArray.vector.index
    DataArray.vector.bounds
-   DataArray.vector.total_bounds
+   DataArray.vector.size
    DataArray.vector.sindex
    DataArray.vector.geometry
+   Dataset.vector.attrs
+   Dataset.vector.crs
+   Dataset.vector.index_dim
+   Dataset.vector.time_dim
+   Dataset.vector.x_name
+   Dataset.vector.y_name
+   Dataset.vector.geom_name
+   Dataset.vector.geom_type
+   Dataset.vector.geom_format
+   Dataset.vector.index
+   Dataset.vector.bounds
+   Dataset.vector.size
+   Dataset.vector.sindex
+   Dataset.vector.geometry
+
+Conversion
+==========
+
+.. autosummary::
+   :template: autosummary/accessor_method.rst
+   :toctree: _generated  
+
+   DataArray.vector.ogr_compliant
+   DataArray.vector.update_geometry
+   DataArray.vector.to_geom
+   DataArray.vector.to_xy
+   DataArray.vector.to_wkt
+   Dataset.vector.ogr_compliant
+   Dataset.vector.update_geometry
+   Dataset.vector.to_geom
+   Dataset.vector.to_xy
+   Dataset.vector.to_wkt
 
 General methods
 ===============
@@ -630,8 +667,8 @@ General methods
 
    DataArray.vector.set_crs
    DataArray.vector.set_spatial_dims
-   DataArray.vector.reset_spatial_dims_attrs
-
+   Dataset.vector.set_crs
+   Dataset.vector.set_spatial_dims
 
 Clip
 ====
@@ -642,6 +679,8 @@ Clip
 
    DataArray.vector.clip_bbox
    DataArray.vector.clip_geom
+   Dataset.vector.clip_bbox
+   Dataset.vector.clip_geom
 
 Reproject
 =========
@@ -651,6 +690,7 @@ Reproject
    :toctree: _generated
 
    DataArray.vector.to_crs
+   Dataset.vector.to_crs
 
 ======================
 Flow direction methods
