### 2023-02-10 6d0d484
small update model build user guide
