+++ docs/conf.py
@@ -355,7 +355,7 @@ intersphinx_mapping = {
     # "dask": ("https://docs.dask.org/en/latest", None),
     "rasterio": ("https://rasterio.readthedocs.io/en/latest", None),
     "geopandas": ("https://geopandas.org/en/stable", None),
-    "xarray": ("https://xarray.pydata.dev/en/stable", None),
+    "xarray": ("https://docs.xarray.dev/en/stable", None),
 }
 
 # -- NBSPHINX --------------------------------------------------------------
