+++ envs/hydromt-dev.yml
@@ -16,6 +16,7 @@ dependencies:
   - descartes # to run examples
   - entrypoints
   - flit>=3.4
+  - fsspec
   - gcsfs
   - gdal>=3.1
   - geopandas>=0.10
@@ -47,6 +48,7 @@ dependencies:
   - universal_pathlib
   - shapely>=2.0.0
   - sphinx # docs
+  - s3fs
   - xarray
   - xugrid>=0.1.5 # optional for mesh
   - zarr
