+++ envs/hydromt-min.yml
@@ -11,6 +11,7 @@ dependencies:
   - dask
   - entrypoints
   - flit>=3.4  # install
+  - fsspec
   - gdal>=3.1
   - geopandas>=0.10
   - netcdf4
