+++ pyproject.toml
@@ -14,6 +14,7 @@ dependencies = [
     "click",
     "dask",
     "entrypoints",
+    "fsspec",
     "geopandas>=0.10",
     "gdal>=3.1",
     "numpy",
