+++ data/catalogs/changelog.rst
@@ -5,12 +5,13 @@ Change log predefined datasets
 deltares_data 
 =============
 
-version: 2022.10
+version: 2023.2
 ----------------
 
 changed
 ^^^^^^^
 - convert GeoPackage files to FlatGeoBuf for cloud compatibility
+- fix ERA5 nc files to read from archive of combined yearly and monthly files
 
 
 version: 2022.7
