+++ data/predefined_catalogs.yml
@@ -8,7 +8,7 @@
 deltares_data:
    urlpath: https://raw.githubusercontent.com/Deltares/hydromt/{version}/data/catalogs/deltares_data.yml
    versions:
-     v2022.10: main
+     v2023.2: main
      v2022.7: e082da339f22cb1fc3571eec5a901a21d1c8a7bd
      v2022.5: d88cc47bd4ecc83de38c00aa554a7d48ad23ec23
    notes: This data is only accessible when connected to the Deltares network.
