### 2023-02-10 a1d8888
add pyet to optional req; fix clip_bbox for rotated grid
