+++ pyproject.toml
@@ -49,6 +49,7 @@ all = [
     "pcraster",
     "xugrid",
     "pygeos",
+    "pyet",
 ]
 test = [
     "black",
