+++ examples/config_file_tests/example_toml.toml
@@ -1,0 +1,11 @@
+[setup_dep]
+[[setup_dep.datasets_dep]]
+dep_fn = "merit_hydro"
+zmax = 10
+zmin = -5
+
+[[setup_dep.datasets_dep]]
+dep_fn = "gebco"
+offset = 1
+zmax = 0
+
