+++ examples/config_file_tests/test_ini.ini
@@ -1,0 +1,2 @@
+[setup_dep]
+datasets_dep = [{"dep_fn":"merit_hydro", "zmax":10, "zmin":-5}, {"dep_fn": "gebco", "offset":1, "zmax":0,}]
