+++ examples/config_file_tests/wflow_build.ini
@@ -1,0 +1,85 @@
+[global]
+data_libs       = []            # add optional paths to data yml files
+
+[setup_config]                  # options parsed to wflow ini file <section>.<option>
+starttime = 2010-01-01T00:00:00
+endtime = 2010-03-31T00:00:00
+timestepsecs = 86400
+input.path_forcing = inmaps-era5-2010.nc
+
+[setup_basemaps]
+hydrography_fn     = merit_hydro   # source hydrography data {merit_hydro, merit_hydro_1k}
+basin_index_fn  = merit_hydro_index # source of basin index corresponding to hydrography_fn
+upscale_method  = ihu           # upscaling method for flow direction data, by default 'ihu'
+
+[setup_rivers]
+hydrography_fn   = merit_hydro      # source hydrography data, should correspond to hydrography_fn in setup_basemaps
+river_geom_fn    = rivers_lin2019_v1 # river source data with river width and bankfull discharge
+river_upa        = 30               # minimum upstream area threshold for the river map [km2]
+rivdph_method    = powlaw           # method to estimate depth {'powlaw', 'manning', 'gvf'}
+min_rivdph       = 1                # minimum river depth [m]
+min_rivwth       = 30               # minimum river width [m]
+slope_len        = 2000             # length over which tp calculate river slope [m]
+smooth_len       = 5000             # length over which to smooth river depth and river width [m]
+
+# [setup_hydrodem]                  # uncomment for river and/or land local-inertial routing.
+# elevtn_map      = dem_subgrid     # {'dem_subgrid', 'wflow_dem'}
+# river_routing   = local-inertial  # {'kinematic-wave', 'local-inertial'}
+# land_routing    = kinematic-wave  # {'kinematic-wave', 'local-inertial'}
+
+[setup_reservoirs]
+reservoirs_fn   = hydro_reservoirs  # source for reservoirs based on GRAND: {hydro_reservoirs}
+min_area        = 1.0           # minimum lake area to consider [km2]
+priority_jrc    = True          # if True then JRC data from hydroengine is used to calculate some reservoir attributes instead of the GRanD and HydroLAKES db.
+
+[setup_lakes]
+lakes_fn        = hydro_lakes   # source for lakes based on hydroLAKES: {hydro_lakes}
+min_area        = 10.0          # minimum reservoir area to consider [km2]
+
+[setup_glaciers]
+glaciers_fn     = rgi           # source for glaciers based on Randolph Glacier Inventory {rgi}
+min_area        = 1.0           # minimum glacier area to consider [km2]
+
+[setup_lulcmaps]
+lulc_fn         = globcover     # source for lulc maps: {globcover, vito, corine}
+
+[setup_laimaps]
+lai_fn          = modis_lai     # source for LAI: {modis_lai}
+
+[setup_soilmaps]
+soil_fn         = soilgrids     # source for soilmaps: {soilgrids}
+ptf_ksatver     = brakensiek    # pedotransfer function to calculate hydraulic conductivity: {brakensiek, cosby}
+
+[setup_gauges]
+gauges_fn       = grdc          # if not None add gaugemap. Either a path or known gauges_fn: {grdc}
+snap_to_river   = True          # if True snaps gauges from source to river
+derive_subcatch = False         # if True derive subcatch map based on gauges.
+
+[setup_precip_forcing]
+precip_fn       = era5_daily_zarr          # source for precipitation.
+precip_clim_fn  = None          # source for high resolution climatology to correct precipitation.
+
+[setup_temp_pet_forcing]
[…]
