### 2023-02-10 c8c68cc
fix renaming bug in GeoDataArray.from_gdf; clear output ipynb
