### 2023-02-10 2b44d42
small edits docs and test; rename cmip6_data to gcs_cmip6_data
