+++ tests/test_data_adapter.py
@@ -8,10 +8,8 @@ import geopandas as gpd
 import pandas as pd
 import xarray as xr
 import hydromt
-
-from hydromt.data_catalog import (
-    DataCatalog,
-)
+from hydromt import _compat as compat
+from hydromt.data_catalog import DataCatalog
 
 TESTDATADIR = join(dirname(abspath(__file__)), "data")
 CATALOGDIR = join(dirname(abspath(__file__)), "..", "data", "catalogs")
@@ -66,9 +64,10 @@ def test_rasterdataset(rioda, tmpdir):
         data_catalog.get_rasterdataset("no_file.tif")
 
 
+@pytest.mark.skipif(not compat.HAS_GCSFS, reason="GCSFS not installed.")
 def test_gcs_cmip6(tmpdir):
     # TODO switch to pre-defined catalogs when pushed to main
-    catalog_fn = join(CATALOGDIR, "cmip6_data.yml")
+    catalog_fn = join(CATALOGDIR, "gcs_cmip6_data.yml")
     data_catalog = DataCatalog(data_libs=[catalog_fn])
     ds = data_catalog.get_rasterdataset(
         "cmip6_NOAA-GFDL/GFDL-ESM4_historical_r1i1p1f1_Amon",
@@ -85,6 +84,7 @@ def test_gcs_cmip6(tmpdir):
     assert np.allclose(ds["precip"][0, :, :], ds1["precip"][0, :, :])
 
 
+@pytest.mark.skipif(not compat.HAS_S3FS, reason="S3FS not installed.")
 def test_aws_copdem(tmpdir):
     # TODO switch to pre-defined catalogs when pushed to main
     catalog_fn = join(CATALOGDIR, "aws_data.yml")
