+++ docs/conf.py
@@ -114,13 +114,18 @@ categories = [
     "ocean",
     "socio-economic",
     "topography",
+    "climate",
     "other",
 ]
 data_cat = hydromt.DataCatalog()
 data_cat.set_predefined_catalogs(r"../data/predefined_catalogs.yml")
 predefined_catalogs = data_cat.predefined_catalogs
 for name in predefined_catalogs:
-    data_cat.from_predefined_catalogs(name)
+    try:
+        data_cat.from_predefined_catalogs(name)
+    except OSError as e:
+        print(e)
+        continue
     note = predefined_catalogs[name].get("notes", "")
     write_nested_dropdown(name, data_cat, note=note, categories=categories)
     data_cat._sources = {}  # reset
