+++ hydromt/data_adapter/dataframe.py
@@ -26,7 +26,6 @@ class DataFrameAdapter(DataAdapter):
         unit_mult={},
         unit_add={},
         meta={},
-        placeholders={},
         **kwargs,
     ):
         """Initiates data adapter for 2D tabular data.
@@ -70,7 +69,6 @@ class DataFrameAdapter(DataAdapter):
             unit_mult=unit_mult,
             unit_add=unit_add,
             meta=meta,
-            placeholders=placeholders,
             **kwargs,
         )
 
