+++ hydromt/data_adapter/rasterdataset.py
@@ -9,6 +9,7 @@ import warnings
 import logging
 from typing import Union, NewType
 from pathlib import Path
+from pyproj import CRS
 
 from .. import gis_utils, io
 from .data_adapter import DataAdapter, PREPROCESSORS
@@ -40,7 +41,6 @@ class RasterDatasetAdapter(DataAdapter):
         unit_add={},
         units={},
         meta={},
-        placeholders={},
         **kwargs,
     ):
         """Initiates data adapter for geospatial raster data.
@@ -75,8 +75,6 @@ class RasterDatasetAdapter(DataAdapter):
         meta: dict, optional
             Metadata information of dataset, preferably containing the following keys:
             {'source_version', 'source_url', 'source_license', 'paper_ref', 'paper_doi', 'category'}
-        placeholders: dict, optional
-            Placeholders to expand yaml entry to multiple entries (name and path) based on placeholder values
         **kwargs
             Additional key-word arguments passed to the driver.
         """
@@ -89,7 +87,6 @@ class RasterDatasetAdapter(DataAdapter):
             unit_mult=unit_mult,
             unit_add=unit_add,
             meta=meta,
-            placeholders=placeholders,
             **kwargs,
         )
         # TODO: see if the units argument can be solved with unit_mult/unit_add
@@ -174,12 +171,68 @@ class RasterDatasetAdapter(DataAdapter):
 
         return fn_out, driver
 
+    def _get_zoom_level(
+        self, zoom_res: float | tuple = None, lat: float = 0, logger=logger
+    ) -> int:
+        """Return nearest smaller zoom level based on zoom resolutions defined in data catalog"""
+        # common pyproj crs axis units
+        known_units = ["degree", "metre", "US survey foot"]
+        if self.zoom_levels is None or len(self.zoom_levels) == 0:
+            logger.warning(f"No zoom levels available, default to zero")
+            return 0
+        zls = list(self.zoom_levels.keys())
+        if zoom_res is None:  # return first zoomlevel (assume these are ordered)
+            return next(iter(zls))
+        # parse zoom_res argument
+        unit = "metre"
+        if (
+            isinstance(zoom_res, tuple)
+            and isinstance(zoom_res[0], (int, float))
+            and isinstance(zoom_res[1], str)
+            and len(zoom_res) == 2
+        ):
+            zoom_res, unit = zoom_res
+            # covert 'meter' and foot to official pyproj units
+            unit = {"meter": "metre", "foot": "US survey foot"}.get(unit, unit)
+            if unit not in known_units:
+                raise TypeError(
+                    f"zoom_res unit {unit} not understood; should be one of {known_units}"
+                )
+        elif not isinstance(zoom_res, (int, float)):
+            raise TypeError(
+                f"zoom_res argument not understood: {zoom_res}; should be a float"
+            )
+        # convert zoom_res if different unit than crs
+        res = zoom_res
+        if self.crs:
+            crs = CRS.from_user_input(self.crs)
+            crs_unit = crs.axis_info[0].unit_name
+            if crs_unit != unit and crs_unit not in known_units:
+                raise NotImplementedError(
+                    f"no conversion available for {unit} to {crs_unit}"
+                )
+            if unit != crs_unit:
+                conversions = {
+                    "degree": np.hypot(*gis_utils.cellres(lat=lat)),
+                    "US survey foot": 0.3048,
+                }
+                res = res * conversions.get(unit, 1) / conversions.get(crs_unit, 1)
+        # find nearest smaller zoomlevel
+        eps = 1e-5  # allow for rounding errors
+        smaller = [x < (res + eps) for x in self.zoom_levels.values()]
[…]
