+++ hydromt/data_catalog.py
@@ -104,7 +104,7 @@ class DataCatalog(object):
     @property
     def keys(self) -> List:
         """Returns list of data source names."""
-        return list(self.sources.keys())
+        return list(self._sources.keys())
 
     @property
     def predefined_catalogs(self) -> Dict:
@@ -113,7 +113,7 @@ class DataCatalog(object):
         return self._catalogs
 
     def __getitem__(self, key: str) -> DataAdapter:
-        return self.sources[key]
+        return self._sources[key]
 
     def __setitem__(self, key: str, value: DataAdapter) -> None:
         if not isinstance(value, DataAdapter):
@@ -123,10 +123,10 @@ class DataCatalog(object):
         return self._sources.__setitem__(key, value)
 
     def __iter__(self):
-        return self.sources.__iter__()
+        return self._sources.__iter__()
 
     def __len__(self):
-        return self.sources.__len__()
+        return self._sources.__len__()
 
     def __repr__(self):
         return self.to_dataframe().__repr__()
@@ -142,6 +142,7 @@ class DataCatalog(object):
     def set_predefined_catalogs(self, urlpath: Union[Path, str] = None) -> Dict:
         # get predefined_catalogs
         urlpath = self._url if urlpath is None else urlpath
+        # TODO cache predefined catalogs
         self._catalogs = _yml_from_uri_or_path(urlpath)
         return self._catalogs
 
@@ -265,6 +266,9 @@ class DataCatalog(object):
               placeholders:
                 <placeholder_name_1>: <list of names>
                 <placeholder_name_2>: <list of names>
+              zoom_levels:
+                <zoom_level_1>: <resolution_1>
+                <zoom_level_2>: <resolution_2>
         """
         self.logger.info(f"Parsing data catalog from {urlpath}")
         yml = _yml_from_uri_or_path(urlpath)
@@ -325,6 +329,7 @@ class DataCatalog(object):
                     "unit_mult": {<hydromt_variable_name1>: <float/int>},
                     "meta": {...},
                     "placeholders": {<placeholder_name_1>: <list of names>},
+                    "zoom_levels": {<zoom_level_1>: <resolution_1>},
                 }
                 <name2>: {
                     ...
@@ -402,7 +407,7 @@ class DataCatalog(object):
             root = abspath(root)
             meta.update(**{"root": root})
             root_drive = os.path.splitdrive(root)[0]
-        for name, source in sorted(self.sources.items()):  # alphabetical order
+        for name, source in sorted(self._sources.items()):  # alphabetical order
             if source_names is not None and name not in source_names:
                 continue
             source_dict = source.to_dict()
@@ -426,7 +431,7 @@ class DataCatalog(object):
     def to_dataframe(self, source_names: List = []) -> pd.DataFrame:
         """Return data catalog summary as DataFrame"""
         d = dict()
-        for name, source in self.sources.items():
+        for name, source in self._sources.items():
             if len(source_names) > 0 and name not in source_names:
                 continue
             d[name] = source.summary()
@@ -464,7 +469,7 @@ class DataCatalog(object):
             os.makedirs(data_root)
 
         # create copy of data with selected source names
-        sources = copy.deepcopy(self.sources)
+        sources = copy.deepcopy(self._sources)
         if len(source_names) > 0:
             sources = {n: sources[n] for n in source_names}
 
@@ -516,6 +521,7 @@ class DataCatalog(object):
         bbox: List = None,
         geom: gpd.GeoDataFrame = None,
         zoom_level: int = None,
+        zoom_res: float = None,
         buffer: Union[float, int] = 0,
         align: bool = None,
         variables: Union[List, str] = None,
@@ -582,6 +588,7 @@ class DataCatalog(object):
             geom=geom,
             buffer=buffer,
[…]
