### 2023-02-12 fa217f1
refactor to single zoom_level argument
