+++ hydromt/data_adapter/data_adapter.py
@@ -3,14 +3,20 @@ from abc import ABCMeta, abstractmethod
 import numpy as np
 import pandas as pd
 import xarray as xr
+import geopandas as gpd
 import yaml
 from upath import UPath
 from fsspec.implementations import local
 from string import Formatter
 from typing import Tuple
 from itertools import product
+from pyproj import CRS
+import logging
 
-from .. import _compat
+from .. import _compat, gis_utils
+
+logger = logging.getLogger(__name__)
+
 
 __all__ = [
     "DataAdapter",
@@ -162,11 +168,76 @@ class DataAdapter(object, metaclass=ABCMeta):
     def __repr__(self):
         return self.__str__()
 
+    def _parse_zoom_level(
+        self,
+        zoom_level: int | tuple = None,
+        geom: gpd.GeoSeries = None,
+        bbox: list = None,
+        logger=logger,
+    ) -> int:
+        """Return nearest smaller zoom level based on zoom resolutions defined in data catalog"""
+        # common pyproj crs axis units
+        known_units = ["degree", "metre", "US survey foot"]
+        if self.zoom_levels is None or len(self.zoom_levels) == 0:
+            logger.warning(f"No zoom levels available, default to zero")
+            return 0
+        zls = list(self.zoom_levels.keys())
+        if zoom_level is None:  # return first zoomlevel (assume these are ordered)
+            return next(iter(zls))
+        # parse zoom_level argument
+        if (
+            isinstance(zoom_level, tuple)
+            and isinstance(zoom_level[0], (int, float))
+            and isinstance(zoom_level[1], str)
+            and len(zoom_level) == 2
+        ):
+            res, unit = zoom_level
+            # covert 'meter' and foot to official pyproj units
+            unit = {"meter": "metre", "foot": "US survey foot"}.get(unit, unit)
+            if unit not in known_units:
+                raise TypeError(
+                    f"zoom_level unit {unit} not understood; should be one of {known_units}"
+                )
+        elif not isinstance(zoom_level, int):
+            raise TypeError(
+                f"zoom_level argument not understood: {zoom_level}; should be a float"
+            )
+        else:
+            return zoom_level
+        if self.crs:
+            # convert res if different unit than crs
+            crs = CRS.from_user_input(self.crs)
+            crs_unit = crs.axis_info[0].unit_name
+            if crs_unit != unit and crs_unit not in known_units:
+                raise NotImplementedError(
+                    f"no conversion available for {unit} to {crs_unit}"
+                )
+            if unit != crs_unit:
+                lat = 0
+                if bbox is not None:
+                    lat = (bbox[1] + bbox[3]) / 2
+                elif geom is not None:
+                    lat = geom.to_crs(4326).centroid.y.item()
+                conversions = {
+                    "degree": np.hypot(*gis_utils.cellres(lat=lat)),
+                    "US survey foot": 0.3048,
+                }
+                res = res * conversions.get(unit, 1) / conversions.get(crs_unit, 1)
+        # find nearest smaller zoomlevel
+        eps = 1e-5  # allow for rounding errors
+        smaller = [x < (res + eps) for x in self.zoom_levels.values()]
+        zl = zls[-1] if all(smaller) else zls[max(smaller.index(False) - 1, 0)]
+        logger.info(f"Getting data for zoom_level {zl} based on res {zoom_level}")
+        return zl
+
     def resolve_paths(
         self,
-        time_tuple: Tuple = None,
-        zoom_level: int = None,
+        time_tuple: tuple = None,
         variables: list = None,
+        zoom_level: int | tuple = None,
+        geom: gpd.GeoSeries = None,
+        bbox: list = None,
+        logger=logger,
         **kwargs,
     ):
[…]
