### 2023-02-12 724f044
fixing minor issues; add exampe to docs
