### 2023-02-12 10b203f
Made create_vrt more flexible
