### 2023-02-12 afafa05
Added kwargs to 'create_vrt'
