### 2023-02-12 1217382
bugfix gdalbuildvrt; annotations py3.10 style
