### 2023-02-14 c529453
make hydromt-min compliant with pyproject.toml
