### 2023-02-14 05dddea
fix name error in GeoDataArray.from_gdf
