+++ hydromt/raster.py
@@ -1964,113 +1964,115 @@ class RasterDataArray(XRasterBase):
         with open(join(root, "..", f"{mName}.yml"), "w") as f:
             yaml.dump({mName: yml}, f, default_flow_style=False, sort_keys=False)
 
-    # def to_osm(
-    #     self,
-    #     root: str,
-    #     zl: int,
-    #     bbox: tuple = (),
-    # ):
-    #     """Generate tiles from raster according to the osm scheme
+    def to_osm(
+        self,
+        root: str,
+        zl: int,
+        bbox: tuple = (),
+    ):
+        """Generate tiles from raster according to the osm scheme
 
-    #     Parameters
-    #     ----------
-    #     root : str
-    #         Path to folder where the database will be created
-    #     zl : int
-    #         Maximum zoom level of the database
-    #         Everyting is generated incrementally up until this level
-    #         E.g. zl = 8, levels generated: 0 to 7
-    #     bbox : tuple, optional
-    #         Bounding Box in the objects crs
+        Parameters
+        ----------
+        root : str
+            Path to folder where the database will be created
+        zl : int
+            Maximum zoom level of the database
+            Everyting is generated incrementally up until this level
+            E.g. zl = 8, levels generated: 0 to 7
+        bbox : tuple, optional
+            Bounding Box in the objects crs
 
-    #     """
+        """
 
-    #     assert self._obj.ndim == 2, "Only 2d datasets are accepted..."
-    #     obj = self._obj.transpose(self.y_dim, self.x_dim)
+        assert self._obj.ndim == 2, "Only 2d datasets are accepted..."
+        obj = self._obj.copy()
+        obj = obj.transpose(self.y_dim, self.x_dim)
+        obj = obj.raster.reproject(dst_crs=3857)
 
-    #     mName = os.path.normpath(os.path.basename(root))
+        mName = os.path.normpath(os.path.basename(root))
 
-    #     def create_folder(path):
-    #         if not os.path.exists(path):
-    #             os.makedirs(path)
+        def create_folder(path):
+            if not os.path.exists(path):
+                os.makedirs(path)
 
-    #     def transform_res(dres, transformer):
-    #         return transformer.transform(0, dres)[0]
+        def transform_res(dres, transformer):
+            return transformer.transform(0, dres)[0]
 
-    #     create_folder(root)
+        create_folder(root)
 
-    #     dres = abs(self._obj.raster.res[0])
-    #     if bbox:
-    #         minx, miny, maxx, maxy = bbox
-    #     else:
-    #         minx, miny, maxx, maxy = self._obj.raster.transform_bounds(
-    #             dst_crs=self._obj.raster.crs
-    #         )
+        dres = abs(self._obj.raster.res[0])
+        if bbox:
+            minx, miny, maxx, maxy = bbox
+        else:
+            minx, miny, maxx, maxy = self._obj.raster.transform_bounds(
+                dst_crs=self._obj.raster.crs
+            )
 
-    #     transformer = pyproj.Transformer.from_crs(self._obj.raster.crs.to_epsg(), 3857)
-    #     minx, miny = map(
-    #         max, zip(transformer.transform(miny, minx), [-20037508.34] * 2)
-    #     )
-    #     maxx, maxy = map(min, zip(transformer.transform(maxy, maxx), [20037508.34] * 2))
+        transformer = pyproj.Transformer.from_crs(self._obj.raster.crs.to_epsg(), 3857)
+        minx, miny = map(
+            max, zip(transformer.transform(miny, minx), [-20037508.34] * 2)
[…]
