### 2023-02-22 f8a7811
Made arguments consistent between raster.clip_geom and raster.clip_mask and updated tests (#284)
