### 2023-02-23 #4897 @DirkEilander
We have just released [v0.7.0](https://github.com/Deltares/hydromt/releases/tag/v0.7.0) which is available on PyPi and conda-forge. This version is also archived at zenodo [DOI:10.5281/zenodo.7663065](https://doi.org/10.5281/zenodo.7663065)

@elbeejay I think the above items have been completed. Please let me know if I'm still missing something.
