### 2023-02-23 453424e
cache predefined data catalogs file to be able to work offline
