### 2023-02-23 3037842
add options to flwdir_from_dem
