### 2023-02-23 #4897 @elbeejay
@DirkEilander there is one outstanding issue you'll need to remedy.

Both the title and list of authors are inconsistent between the `paper.md` file (which gets rendered into the JOSS paper like above) and the Zenodo archive you provided. We need both of these to match before we can proceed.

Thanks!
