### 2023-02-24 7dc4664
Added parse_toml_config function to config.py and updated tests (#252)
