+++ tests/test_config.py
@@ -1,13 +1,14 @@
 import pytest
 from pathlib import Path
 from hydromt import config
+import toml
 
 
 def test_config(tmpdir):
     cfdict = {
         "section1": {
             "list": [1, 2, 3],
-            "tuple": (1, "b"),
+            # "tuple": (1, "b"), seems unnecessary to test for tuple, testing parsing toml will fail if included
             "bool": True,
             "str": "test",
             "int": 1,
@@ -25,7 +26,13 @@ def test_config(tmpdir):
         },
     }
     config_fn = tmpdir.join("config.ini")
+    config_fn2 = tmpdir.join("config.toml")
+    with open(config_fn2, "w") as toml_file:
+        toml.dump(cfdict, toml_file)
     config.configwrite(config_fn, cfdict)
+    # test for deprecation warning
+    with pytest.deprecated_call():
+        config.configread(config_fn=config_fn)
     cfdict1 = config.configread(config_fn, abs_path=True)
     assert cfdict["section1"] == cfdict1["section1"]
     assert isinstance(cfdict1["section2"]["path"], Path)
@@ -40,3 +47,13 @@ def test_config(tmpdir):
     # do not evaluate a specific section
     cfdict1 = config.configread(config_fn, skip_eval_sections=["setup_config"])
     assert isinstance(cfdict1["setup_config"]["float"], str)
+    cfdict2 = config.configread(config_fn2, abs_path=True)
+    cfdict["section1"].pop(
+        "None", None
+    )  # None is dropped from the dictionary when writing to toml
+    assert cfdict["section1"] == cfdict2["section1"]
+    assert isinstance(cfdict2["section2"]["path"], Path)
+    assert isinstance(cfdict2["section2"]["path1"], str)
+    # by default paths in setup_config are not evaluated
+    assert isinstance(cfdict2["setup_config"]["path"], str)
+    assert isinstance(cfdict2["setup_config"]["float"], float)
