+++ examples/config_file_tests/toml_test.toml
@@ -1,0 +1,15 @@
+[section1]
+list = [ 1, 2, 3,]
+tuple = [ 1, "b",]
+bool = true
+str = "test"
+int = 1
+float = 2.3
+
+[section2]
+path = "config.ini"
+path1 = "config1.ini"
+
+[setup_config]
+path = "config.ini"
+float = 2.3
