+++ paper.md
@@ -23,9 +23,9 @@ authors:
   - name: Anaïs Couasnon
     orcid: 0000-0001-9372-841X
     affiliation: "1, 2"
-  - name: Mark Hegnauer
-    affiliation: 1
   - name: Brendam Dalmijn
+    affiliation: 1
+  - name: Mark Hegnauer
     affiliation: 1
   - name: Tjalling de Jong
     affiliation: 1
