### 2023-03-07 #4897 @DirkEilander
@elbeejay I've updated the authors list (with their permissions) in the paper.md manuscript and aligned it with zenodo. As a result I also had to change the acknowledgements. Hope it's in line with the requirements now. Let me know if you have remaining questions.
